refactor(NANO-113): decompose GUIServer into domain handler modules
Split server.py (5,548 lines, 77 handlers) into 8 domain modules:
- server_memory.py (12 handlers): memory config, CRUD, search, promote
- server_sessions.py (7 handlers): session list, resume, create, delete
- server_vts.py (7 handlers): VTubeStudio config and control
- server_stimuli.py (6 handlers): stimuli config, twitch, addressing-others
- server_config.py (11 handlers): personas, VAD, pipeline, blocks, gen params
- server_providers.py (11 handlers): LLM/VLM/tools config, launch, swap
- server_avatar.py (9 handlers): avatar config, Tauri install, expressions

server.py retains 14 core handlers (connect/disconnect, lifecycle,
text input, state/health/config) + all emit_* public API methods.

Pattern: standalone registration functions receive GUIServer instance,
register closures on sio. Zero behavior changes — mechanical self→server
transformation only.

Tests updated: _extract_model_name, _validate_local_*_config,
_persist_local_vlm_config, _emit_sessions mock references all point
to new module locations.

1,950 backend tests passed, 448 frontend tests passed, next build clean.
Verified live: STT OFF + VLM OFF config, text input, Twitch stimuli,
patience timer, LLM provider hot-swap, generation param changes,
avatar/subtitle spawn, memory RAG injection.
