+++ src/spindl/gui/server.py
@@ -31,6 +31,17 @@ import socketio
 
 from spindl.characters.loader import CharacterLoader
 from spindl.launcher.service_runner import kill_process_tree
+from spindl.gui.server_memory import register_memory_handlers
+from spindl.gui.server_sessions import register_session_handlers, emit_sessions
+from spindl.gui.server_vts import register_vts_handlers
+from spindl.gui.server_stimuli import register_stimuli_handlers, build_stimuli_hydration
+from spindl.gui.server_config import register_config_handlers, emit_personas
+from spindl.gui.server_providers import (
+    register_provider_handlers, get_llm_provider_info, get_vlm_provider_info,
+)
+from spindl.gui.server_avatar import (
+    register_avatar_handlers, tauri_binary_path, tauri_binary_exists,
+)
 
 if TYPE_CHECKING:
     from spindl.orchestrator.voice_agent import VoiceAgentOrchestrator
@@ -204,357 +215,13 @@ class GUIServer:
             if stimuli_section and isinstance(stimuli_section, dict):
                 from ..orchestrator.config import StimuliConfig
                 parsed = StimuliConfig.from_dict(stimuli_section)
-                self._stimuli_config_cache = GUIServer._build_stimuli_hydration(parsed)
+                self._stimuli_config_cache = build_stimuli_hydration(parsed)
 
             print(f"[GUI] Pre-launch paths: conversations={self._conversations_dir}, personas={self._personas_dir}", flush=True)
 
         except Exception as e:
             print(f"[GUI] Failed to parse config paths for pre-launch access: {e}", flush=True)
 
-    def _get_block_config_pre_launch(self) -> dict:
-        """
-        Build block config from YAML + defaults without orchestrator (NANO-064b).
-
-        Mirrors VoiceAgentOrchestrator.get_block_config() serialization.
-        """
-        from spindl.llm.prompt_block import create_default_blocks, load_block_config
-
-        if self._prompt_blocks_config:
-            blocks = load_block_config(self._prompt_blocks_config)
-        else:
-            blocks = create_default_blocks()
-
-        block_infos = []
-        order_list = []
-        disabled_list = []
-        overrides_dict: dict[str, str | None] = {}
-
-        for block in sorted(blocks, key=lambda b: b.order):
-            order_list.append(block.id)
-            if not block.enabled:
-                disabled_list.append(block.id)
-            if block.user_override is not None:
-                overrides_dict[block.id] = block.user_override
-
-            block_infos.append({
-                "id": block.id,
-                "label": block.label,
-                "order": block.order,
-                "enabled": block.enabled,
-                "is_static": block.is_static,
-                "section_header": block.section_header,
-                "has_override": block.user_override is not None,
-                "content_wrapper": block.content_wrapper,
-            })
-
-        return {
-            "order": order_list,
-            "disabled": disabled_list,
-            "overrides": overrides_dict,
-            "blocks": block_infos,
-        }
-
-    def _save_block_config_pre_launch(self, config: dict) -> bool:
-        """
-        Persist prompt_blocks config to YAML without orchestrator (NANO-064b).
-
-        Uses ruamel.yaml round-trip (NANO-106).
-
-        Returns:
-            True if persisted successfully, False otherwise.
-        """
-        if not self._config_path:
-            return False
-
-        try:
-            from spindl.orchestrator.config import _make_ruamel_yaml
-
-            config_path = Path(self._config_path)
-            if not config_path.exists():
-                return False
-
-            has_content = bool(
-                config.get("order") or config.get("disabled")
-                or config.get("overrides") or config.get("wrappers")
-            )
-
-            ry = _make_ruamel_yaml()
-            with open(config_path, "r", encoding="utf-8") as f:
-                data = ry.load(f)
-
-            if has_content:
-                data["prompt_blocks"] = config
-            else:
-                if "prompt_blocks" in data:
-                    del data["prompt_blocks"]
-
-            with open(config_path, "w", encoding="utf-8") as f:
-                ry.dump(data, f)
-
-            print(f"[GUI] Block config persisted pre-launch to {config_path.name}", flush=True)
-            return True
-
-        except Exception as e:
-            print(f"[GUI] Failed to persist block config pre-launch: {e}", flush=True)
-            return False
-
-    @staticmethod
-    def _validate_local_llm_config(config: dict) -> list[str]:
-        """Validate local LLM config dict before persisting (NANO-089).
-
-        Checks required fields for a local llama launch config.
-        Returns list of error strings (empty if valid).
-        """
-        import os
-
-        errors = []
-        has_url = bool(config.get("url"))
-        has_local = bool(config.get("executable_path") or config.get("model_path"))
-        if not has_url and not has_local:
-            errors.append("Local LLM config requires executable_path + model_path, or url")
-        if has_local:
-            if not config.get("executable_path"):
-                errors.append("executable_path is required for local LLM launch")
-            if not config.get("model_path"):
-                errors.append("model_path is required for local LLM launch")
-        port = config.get("port")
-        if port is not None and (not isinstance(port, int) or port < 1 or port > 65535):
-            errors.append(f"port must be 1-65535, got {port}")
-
-        # mmproj_path: if set, file must exist (Session 606: stale path = silent launch failure)
-        mmproj_path = config.get("mmproj_path")
-        if mmproj_path and isinstance(mmproj_path, str) and not os.path.isfile(mmproj_path):
-            errors.append(
-                f"mmproj_path does not exist: {mmproj_path} — "
-                "stale path from a previous model? Clear it if VLM is disabled"
-            )
-
-        return errors
-
-    @staticmethod
-    def _strip_empty_optional_fields(config: dict) -> None:
-        """Remove keys whose values are empty strings — prevents zombie fields in YAML.
-
-        Fields like mmproj_path, reasoning_format, extra_args can be "" when the
-        user clears them in the GUI. Leaving "" in YAML means the next .update()
-        merge won't delete them, and the backend may interpret their presence
-        as intentional config even though the value is blank.
-        """
-        zombie_keys = [
-            k for k, v in config.items()
-            if isinstance(v, str) and v == ""
-        ]
-        for k in zombie_keys:
-            del config[k]
-
-    def _persist_local_llm_config(self, config: dict) -> bool:
-        """
-        Persist llama provider config to YAML after successful launch (NANO-065b).
-
-        Updates llm.providers.llama section and sets llm.provider to "llama".
-        Works both with orchestrator (delegates to save_to_yaml) and pre-launch
-        (direct YAML surgery).
-
-        Returns:
-            True if persisted successfully.
-        """
-        # NANO-089: validate before persisting
-        errors = self._validate_local_llm_config(config)
-        if errors:
-            print(f"[GUI] LLM config validation failed: {errors}", flush=True)
-            return False
-
-        # Strip empty-string fields before merge — prevents zombie keys in YAML
-        self._strip_empty_optional_fields(config)
-
-        if self._orchestrator:
-            try:
-                # Update orchestrator config in-memory, then persist.
-                # Merge onto existing to preserve dashboard-only keys
-                # (e.g. repeat_penalty from NANO-108).
-                llm_cfg = self._orchestrator._config.llm_config
-                if "llama" not in llm_cfg.providers:
-                    llm_cfg.providers["llama"] = {}
-                existing = llm_cfg.providers["llama"]
-                # Delete keys that the incoming config explicitly cleared
-                for k in list(existing.keys()):
-                    if k in ("mmproj_path", "reasoning_format", "extra_args", "device", "tensor_split") and k not in config:
-                        del existing[k]
-                existing.update(config)
-                self._orchestrator._config.save_to_yaml(self._config_path)
-                print(
-                    f"[GUI] Local LLM config persisted via orchestrator to "
-                    f"{Path(self._config_path).name}",
-                    flush=True,
-                )
-                return True
-            except Exception as e:
-                print(f"[GUI] Failed to persist local LLM config: {e}", flush=True)
-                return False
-
-        # Pre-launch: ruamel.yaml round-trip (NANO-106)
-        if not self._config_path:
-            return False
-
-        try:
-            from spindl.orchestrator.config import _make_ruamel_yaml
-
-            config_path = Path(self._config_path)
-            if not config_path.exists():
-                return False
-
-            ry = _make_ruamel_yaml()
-            with open(config_path, "r", encoding="utf-8") as f:
-                data = ry.load(f)
-
-            if "llm" not in data:
-                data["llm"] = {}
-            if "providers" not in data["llm"]:
-                data["llm"]["providers"] = {}
-            # Merge launcher config onto existing section to preserve
-            # dashboard-only keys (e.g. repeat_penalty from NANO-108)
-            if "llama" not in data["llm"]["providers"]:
-                data["llm"]["providers"]["llama"] = {}
-            existing_yaml = data["llm"]["providers"]["llama"]
-            # Delete zombie keys that were cleared (not in incoming config)
-            for k in list(existing_yaml.keys()):
-                if k in ("mmproj_path", "reasoning_format", "extra_args", "device", "tensor_split") and k not in config:
-                    del existing_yaml[k]
-            existing_yaml.update(config)
-
-            with open(config_path, "w", encoding="utf-8") as f:
-                ry.dump(data, f)
-
-            # Update cache
-            self._llm_config_cache = self._llm_config_cache or {}
-            if "providers" not in self._llm_config_cache:
-                self._llm_config_cache["providers"] = {}
-            self._llm_config_cache["providers"]["llama"] = config
-
-            print(
-                f"[GUI] Local LLM config persisted pre-launch to {config_path.name}",
-                flush=True,
-            )
-            return True
-
-        except Exception as e:
-            print(f"[GUI] Failed to persist local LLM config pre-launch: {e}", flush=True)
-            return False
-
-    @staticmethod
-    def _validate_local_vlm_config(config: dict) -> list[str]:
-        """Validate local VLM config dict before persisting (NANO-089).
-
-        Checks required fields for a local llama VLM launch config.
-        Returns list of error strings (empty if valid).
-        """
-        import os
-
-        errors = []
-        if not config.get("executable_path"):
-            errors.append("executable_path is required for local VLM launch")
-        if not config.get("model_path"):
-            errors.append("model_path is required for local VLM launch")
-        if not config.get("model_type"):
-            errors.append("model_type is required for local VLM launch")
-        port = config.get("port")
-        if port is not None and (not isinstance(port, int) or port < 1 or port > 65535):
-            errors.append(f"port must be 1-65535, got {port}")
-
-        # mmproj_path: if set, file must exist (Session 606: stale path = silent launch failure)
-        mmproj_path = config.get("mmproj_path")
-        if mmproj_path and isinstance(mmproj_path, str) and not os.path.isfile(mmproj_path):
-            errors.append(
-                f"mmproj_path does not exist: {mmproj_path} — "
-                "stale path from a previous model? Clear it if VLM is disabled"
-            )
-
-        return errors
-
-    def _persist_local_vlm_config(self, config: dict) -> bool:
-        """
-        Persist llama VLM provider config to YAML after successful launch (NANO-079).
-
-        Updates vision.providers.llama section and sets vision.provider to "llama".
-        """
-        # NANO-089: validate before persisting
-        errors = self._validate_local_vlm_config(config)
-        if errors:
-            print(f"[GUI] VLM config validation failed: {errors}", flush=True)
-            return False
-
-        # Strip empty-string fields before merge — prevents zombie keys in YAML
-        self._strip_empty_optional_fields(config)
-
-        if self._orchestrator:
-            try:
-                vis_cfg = self._orchestrator._config.vlm_config
-                vis_cfg.provider = "llama"
-                if "llama" not in vis_cfg.providers:
-                    vis_cfg.providers["llama"] = {}
-                existing = vis_cfg.providers["llama"]
-                # Delete zombie keys that were cleared
-                for k in list(existing.keys()):
-                    if k in ("mmproj_path", "extra_args", "device", "tensor_split") and k not in config:
-                        del existing[k]
-                existing.update(config)
-                self._orchestrator._config.save_to_yaml(self._config_path)
-                print(
-                    f"[GUI] Local VLM config persisted via orchestrator to "
-                    f"{Path(self._config_path).name}",
-                    flush=True,
-                )
-                return True
-            except Exception as e:
-                print(f"[GUI] Failed to persist local VLM config: {e}", flush=True)
-                return False
-
-        # Pre-launch: ruamel.yaml round-trip (NANO-106)
-        if not self._config_path:
-            return False
-
-        try:
-            from spindl.orchestrator.config import _make_ruamel_yaml
-
-            config_path = Path(self._config_path)
-            if not config_path.exists():
-                return False
-
-            ry = _make_ruamel_yaml()
-            with open(config_path, "r", encoding="utf-8") as f:
-                data = ry.load(f)
-
-            if "vlm" not in data:
-                data["vlm"] = {}
-            if "providers" not in data["vlm"]:
-                data["vlm"]["providers"] = {}
-            if "llama" not in data["vlm"]["providers"]:
-                data["vlm"]["providers"]["llama"] = {}
-            existing_yaml = data["vlm"]["providers"]["llama"]
-            # Delete zombie keys that were cleared
-            for k in list(existing_yaml.keys()):
-                if k in ("mmproj_path", "extra_args", "device", "tensor_split") and k not in config:
-                    del existing_yaml[k]
-            existing_yaml.update(config)
-
-            with open(config_path, "w", encoding="utf-8") as f:
-                ry.dump(data, f)
-
-            # Update cache
-            self._vlm_config_cache = self._vlm_config_cache or {}
-            if "providers" not in self._vlm_config_cache:
-                self._vlm_config_cache["providers"] = {}
-            self._vlm_config_cache["providers"]["llama"] = config
-
-            print(
-                f"[GUI] Local VLM config persisted pre-launch to {config_path.name}",
-                flush=True,
-            )
-            return True
-
-        except Exception as e:
-            print(f"[GUI] Failed to persist local VLM config pre-launch: {e}", flush=True)
-            return False
 
     def attach(self, orchestrator: "VoiceAgentOrchestrator") -> None:
         """Attach the orchestrator for event bridging.
@@ -666,6 +333,15 @@ class GUIServer:
     def _register_handlers(self) -> None:
         """Register Socket.IO event handlers."""
 
+        # Domain-specific handler modules (NANO-113)
+        register_memory_handlers(self)
+        register_session_handlers(self)
+        register_config_handlers(self)
+        register_provider_handlers(self)
+        register_stimuli_handlers(self)
+        register_vts_handlers(self)
+        register_avatar_handlers(self)
+
         @self.sio.event
         async def connect(sid: str, environ: dict) -> None:
             self._clients.add(sid)
@@ -704,126 +380,6 @@ class GUIServer:
             print(f"[GUI] Client disconnected (total: {len(self._clients)})", flush=True)
 
         @self.sio.event
-        async def register_avatar_client(sid: str, data: dict) -> None:
-            """Avatar renderer identifies itself on connect (NANO-097)."""
-            self._avatar_clients.add(sid)
-            print(
-                f"[GUI] Avatar client registered "
-                f"(avatar clients: {len(self._avatar_clients)})",
-                flush=True,
-            )
-            await self.sio.emit(
-                "avatar_connection_status",
-                {"connected": True},
-            )
-            # Push active character's VRM + config to the newly connected avatar
-            if self._orchestrator:
-                try:
-                    config = self._orchestrator._config
-                    loader = CharacterLoader(config.characters_dir)
-                    vrm_path = loader.get_vrm_path(config.character_id)
-                    expressions = loader.get_avatar_expressions(config.character_id)
-                    animations = loader.get_avatar_animations(config.character_id)
-                    char_anim_dir = str(loader.get_character_animations_dir(config.character_id))
-                    payload: dict = {"path": str(vrm_path) if vrm_path else ""}
-                    if expressions:
-                        payload["expressions"] = expressions
-                    if animations:
-                        payload["animations"] = animations
-                    if char_anim_dir:
-                        payload["character_animations_dir"] = char_anim_dir
-                    await self.sio.emit(
-                        "avatar_load_model",
-                        payload,
-                        to=sid,
-                    )
-                    print(
-                        f"[GUI] Avatar load_model pushed to new client: {vrm_path or '(default)'}",
-                        flush=True,
-                    )
-                except Exception as e:
-                    print(f"[GUI] Avatar load_model on register failed: {e}", flush=True)
-
-        @self.sio.event
-        async def preview_avatar_expressions(sid: str, data: dict) -> None:
-            """Live preview of expression composites from character editor (NANO-098).
-
-            Relays to all connected avatar clients without persisting.
-            Editor is responsible for reverting on cancel/navigate-away.
-            """
-            expressions = data.get("expressions", {})
-            preview_mood = data.get("previewMood")
-            payload: dict = {"expressions": expressions}
-            if preview_mood is not None:
-                payload["previewMood"] = preview_mood
-            await self.sio.emit("avatar_preview_expressions", payload)
-
-        @self.sio.event
-        async def update_avatar_animation_config(sid: str, data: dict) -> None:
-            """Push updated animation config to renderer on save (NANO-098 Session 3).
-
-            Sent directly from editor with the config data — no disk read needed.
-            Avoids race condition with card.json write.
-            """
-            animations = data.get("animations")
-            await self.sio.emit("avatar_update_animation_config", {"animations": animations})
-
-        @self.sio.event
-        async def preview_avatar_animation(sid: str, data: dict) -> None:
-            """Live preview of animation clip from character editor (NANO-098 Session 3).
-
-            Relays clip name to all connected avatar clients.
-            clip=null means stop playing (revert to procedural idle).
-            """
-            clip = data.get("clip")
-            await self.sio.emit("avatar_preview_animation", {"clip": clip})
-
-        @self.sio.event
-        async def reload_avatar_model(sid: str, data: dict) -> None:
-            """Dashboard requests avatar model reload after VRM change (NANO-097)."""
-            character_id = data.get("character_id")
-            if not character_id or not self._orchestrator:
-                return
-            # Only push if this is the active character
-            if character_id != self._orchestrator._config.character_id:
-                return
-            try:
-                config = self._orchestrator._config
-                loader = CharacterLoader(config.characters_dir)
-                vrm_path = loader.get_vrm_path(character_id)
-                if vrm_path:
-                    expressions = loader.get_avatar_expressions(character_id)
-                    animations = loader.get_avatar_animations(character_id)
-                    char_anim_dir = str(loader.get_character_animations_dir(character_id))
-                    await self.emit_avatar_load_model(
-                        str(vrm_path), expressions=expressions,
-                        animations=animations, character_animations_dir=char_anim_dir,
-                    )
-                    print(
-                        f"[GUI] Avatar model reloaded: {vrm_path}",
-                        flush=True,
-                    )
-                else:
-                    # No VRM — tell avatar to load default model, but still send config
-                    expressions = loader.get_avatar_expressions(config.character_id)
-                    animations = loader.get_avatar_animations(config.character_id)
-                    char_anim_dir = str(loader.get_character_animations_dir(config.character_id))
-                    payload: dict = {"path": ""}
-                    if expressions:
-                        payload["expressions"] = expressions
-                    if animations:
-                        payload["animations"] = animations
-                    if char_anim_dir:
-                        payload["character_animations_dir"] = char_anim_dir
-                    await self.sio.emit("avatar_load_model", payload)
-                    print(
-                        "[GUI] Avatar model reloaded: (default)",
-                        flush=True,
-                    )
-            except Exception as e:
-                print(f"[GUI] Avatar model reload failed: {e}", flush=True)
-
-        @self.sio.event
         async def request_state(sid: str, data: dict) -> None:
             """Client requests current state snapshot."""
             await self._emit_state(sid)
@@ -839,3013 +395,12 @@ class GUIServer:
             await self._emit_config(sid)
 
         @self.sio.event
-        async def request_sessions(sid: str, data: dict) -> None:
-            """Client requests session list."""
-            await self._emit_sessions(sid, data.get("persona"))
-
-        @self.sio.event
-        async def request_session_detail(sid: str, data: dict) -> None:
-            """Client requests session details."""
-            filepath = data.get("filepath")
-            if filepath:
-                await self._emit_session_detail(sid, filepath)
-
-        @self.sio.event
-        async def request_chat_history(sid: str, data: dict) -> None:
-            """Client requests chat history for the active session (NANO-073a)."""
-            from spindl.history import jsonl_store
-
-            if not self._orchestrator or not self._orchestrator.session_file:
-                await self.sio.emit(
-                    "chat_history",
-                    {"turns": []},
-                    to=sid,
-                )
-                return
-
-            try:
-                filepath = Path(self._orchestrator.session_file)
-                if not filepath.exists():
-                    await self.sio.emit(
-                        "chat_history",
-                        {"turns": []},
-                        to=sid,
-                    )
-                    return
-
-                visible = jsonl_store.read_visible_turns(filepath)
-                # Map to frontend-friendly format with metadata (NANO-075)
-                turns = []
-                for t in visible:
-                    role = t.get("role")
-                    if role in ("user", "assistant"):
-                        # NANO-109: display_content holds raw LLM output (with
-                        # formatting); content holds cleaned text for LLM replay.
-                        # Chat display should show the raw version when available.
-                        display_text = t.get("display_content") or t.get("content", "")
-                        turn = {
-                            "role": role,
-                            "text": display_text,
-                            "timestamp": t.get("timestamp", ""),
-                        }
-                        # NANO-075: Forward metadata for hydration survival
-                        if role == "user":
-                            input_mod = t.get("input_modality")
-                            if input_mod:
-                                turn["input_modality"] = input_mod
-                        elif role == "assistant":
-                            reasoning = t.get("reasoning")
-                            if reasoning:
-                                turn["reasoning"] = reasoning
-                            stimulus = t.get("stimulus_source")
-                            if stimulus:
-                                turn["stimulus_source"] = stimulus
-                            codex = t.get("activated_codex_entries")
-                            if codex:
-                                turn["activated_codex_entries"] = codex
-                            memories = t.get("retrieved_memories")
-                            if memories:
-                                turn["retrieved_memories"] = memories
-                            # NANO-094: Emotion classifier metadata
-                            emotion = t.get("emotion")
-                            if emotion:
-                                turn["emotion"] = emotion
-                                turn["emotion_confidence"] = t.get("emotion_confidence")
-                        turns.append(turn)
-
-                # Cap at 200 most recent to avoid DOM bloat
-                if len(turns) > 200:
-                    turns = turns[-200:]
-
-                await self.sio.emit(
-                    "chat_history",
-                    {"turns": turns},
-                    to=sid,
-                )
-            except Exception as e:
-                print(f"[GUI] Error reading chat history: {e}", flush=True)
-                await self.sio.emit(
-                    "chat_history",
-                    {"turns": []},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def request_prompt_snapshot(sid: str, data: dict | None = None) -> None:
-            """Client requests current prompt snapshot (NANO-076)."""
-            if not self._orchestrator:
-                # No orchestrator — frontend stays on currentSnapshot: null
-                return
-
-            try:
-                snapshot = self._orchestrator.get_prompt_snapshot()
-                if snapshot:
-                    await self.sio.emit("prompt_snapshot", snapshot, to=sid)
-                # else: no snapshot available — frontend stays on null (cleared on session switch)
-            except Exception as e:
-                print(f"[GUI] Error building prompt snapshot: {e}", flush=True)
-
-        @self.sio.event
-        async def resume_session(sid: str, data: dict) -> None:
-            """Client requests to resume a specific session by filename."""
-            filename = data.get("filename")
-            if not filename or not self._orchestrator:
-                return
-
-            # Resolve filename to full path in conversations directory
-            conv_dir_str = self._orchestrator._config.conversations_dir
-            if not conv_dir_str:
-                return
-            resolved = Path(conv_dir_str) / filename
-            if not resolved.exists():
-                print(f"[GUI] Session resume failed: {filename} not found", flush=True)
-                await self.sio.emit(
-                    "session_resumed",
-                    {"filepath": str(resolved), "success": False, "error": "Session file not found"},
-                    to=sid,
-                )
-                return
-
-            success = self._orchestrator.load_session(str(resolved))
-            if success:
-                print(f"[GUI] Session resumed: {filename}", flush=True)
-                await self.sio.emit(
-                    "session_resumed",
-                    {"filepath": str(resolved), "success": True},
-                    to=sid,
-                )
-                # Refresh session list so active_session badge updates
-                await self._emit_sessions()
-            else:
-                print(f"[GUI] Session resume failed: {filename}", flush=True)
-                await self.sio.emit(
-                    "session_resumed",
-                    {"filepath": str(resolved), "success": False, "error": "Failed to load session"},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def create_session(sid: str, data: dict | None = None) -> None:
-            """Client requests a new session for the current persona (NANO-071)."""
-            if not self._orchestrator:
-                await self.sio.emit(
-                    "session_created",
-                    {"success": False, "error": "Services not running"},
-                    to=sid,
-                )
-                return
-
-            success = self._orchestrator.create_new_session()
-            if success:
-                new_filepath = str(self._orchestrator.session_file) if self._orchestrator.session_file else None
-                # Touch the file so it exists on disk for _emit_sessions glob
-                if new_filepath:
-                    Path(new_filepath).parent.mkdir(parents=True, exist_ok=True)
-                    Path(new_filepath).touch()
-                print(f"[GUI] New session created: {Path(new_filepath).name if new_filepath else 'unknown'}", flush=True)
-                await self.sio.emit(
-                    "session_created",
-                    {"success": True, "filepath": new_filepath},
-                    to=sid,
-                )
-                # Refresh session list for all clients
-                await self._emit_sessions()
-            else:
-                await self.sio.emit(
-                    "session_created",
-                    {"success": False, "error": "Failed to create new session"},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def delete_session(sid: str, data: dict) -> None:
-            """Client requests to delete a session file (NANO-064a)."""
-            filepath = data.get("filepath")
-            if not filepath:
-                return
-
-            try:
-                path = Path(filepath)
-                if not path.exists():
-                    await self.sio.emit(
-                        "session_deleted",
-                        {"filepath": filepath, "success": False, "error": "Session not found"},
-                        to=sid,
-                    )
-                    return
-
-                # Don't allow deleting the active session (only when services are running)
-                if (
-                    self._orchestrator
-                    and self._orchestrator.session_file
-                    and path == self._orchestrator.session_file
-                ):
-                    await self.sio.emit(
-                        "session_deleted",
-                        {"filepath": filepath, "success": False, "error": "Cannot delete active session"},
-                        to=sid,
-                    )
-                    return
-
-                path.unlink()
-                # NANO-076: Clean up snapshot sidecar if it exists
-                from spindl.history.snapshot_store import delete_sidecar
-                delete_sidecar(path)
-                print(f"[GUI] Session deleted: {path.name}", flush=True)
-                await self.sio.emit(
-                    "session_deleted",
-                    {"filepath": filepath, "success": True},
-                    to=sid,
-                )
-                # Refresh session list for all clients
-                await self._emit_sessions()
-
-            except Exception as e:
-                print(f"[GUI] Session delete error: {e}", flush=True)
-                await self.sio.emit(
-                    "session_deleted",
-                    {"filepath": filepath, "success": False, "error": str(e)},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def generate_session_summary(sid: str, data: dict) -> None:
-            """Client requests session summary generation (NANO-043 Phase 4)."""
-            filepath = data.get("filepath")
-            if filepath and self._orchestrator:
-                try:
-                    # Run in executor to avoid blocking event loop (LLM call)
-                    loop = asyncio.get_event_loop()
-                    summary = await loop.run_in_executor(
-                        None,
-                        self._orchestrator.generate_session_summary,
-                        filepath,
-                    )
-                    session_name = Path(filepath).name
-                    if summary:
-                        print(f"[GUI] Session summary generated: {session_name}", flush=True)
-                    else:
-                        print(f"[GUI] Session summary failed: {session_name}", flush=True)
-                    await self.sio.emit(
-                        "session_summary_generated",
-                        {
-                            "filepath": filepath,
-                            "success": summary is not None,
-                            "summary_preview": summary[:200] if summary else None,
-                            "error": None if summary else "Failed to generate summary",
-                        },
-                        to=sid,
-                    )
-                except Exception as e:
-                    print(f"[GUI] Session summary error: {e}", flush=True)
-                    await self.sio.emit(
-                        "session_summary_generated",
-                        {
-                            "filepath": filepath,
-                            "success": False,
-                            "summary_preview": None,
-                            "error": str(e),
-                        },
-                        to=sid,
-                    )
-
-        @self.sio.event
-        async def request_personas(sid: str, data: dict) -> None:
-            """Client requests available personas list."""
-            await self._emit_personas(sid)
-
-        @self.sio.event
-        async def set_persona(sid: str, data: dict) -> None:
-            """Client requests persona change (NANO-077: runtime hot-swap)."""
-            persona_id = data.get("persona_id")
-            if not persona_id or not self._orchestrator:
-                await self.sio.emit(
-                    "persona_change_failed",
-                    {"error": "Missing persona_id or orchestrator not ready"},
-                    to=sid,
-                )
-                return
-
-            try:
-                success = self._orchestrator.switch_persona(persona_id)
-            except Exception as e:
-                print(f"[GUI] Persona switch failed: {e}", flush=True)
-                await self.sio.emit(
-                    "persona_change_failed",
-                    {"error": str(e)},
-                    to=sid,
-                )
-                return
-
-            if success:
-                print(f"[GUI] Persona switched to: {persona_id}", flush=True)
-
-                # Persist to spindl.yaml
-                if self._config_path:
-                    try:
-                        self._orchestrator._config.save_to_yaml(self._config_path)
-                    except Exception as e:
-                        print(f"[GUI] Failed to persist character.default: {e}", flush=True)
-
-                # Broadcast to ALL clients
-                await self.sio.emit(
-                    "persona_changed",
-                    {"persona_id": persona_id, "restart_required": False},
-                )
-                # Push updated session list (filtered to new character)
-                await self._emit_sessions(sid, persona_id)
-                # Push updated config (persona name/id changed)
-                await self._emit_config(sid)
-
-                # NANO-097: Notify avatar renderer of new character's VRM model
-                if self._orchestrator._config.avatar_config.enabled and self.has_avatar_client:
-                    try:
-                        loader = CharacterLoader(
-                            self._orchestrator._config.characters_dir
-                        )
-                        vrm_path = loader.get_vrm_path(persona_id)
-                        if vrm_path:
-                            expressions = loader.get_avatar_expressions(persona_id)
-                            animations = loader.get_avatar_animations(persona_id)
-                            char_anim_dir = str(loader.get_character_animations_dir(persona_id))
-                            await self.emit_avatar_load_model(
-                                str(vrm_path), expressions=expressions,
-                                animations=animations, character_animations_dir=char_anim_dir,
-                            )
-                            print(
-                                f"[GUI] Avatar load_model emitted: {vrm_path}",
-                                flush=True,
-                            )
-                        else:
-                            # Character has no VRM — tell avatar to load default, but still send config
-                            expressions = loader.get_avatar_expressions(persona_id)
-                            animations = loader.get_avatar_animations(persona_id)
-                            char_anim_dir = str(loader.get_character_animations_dir(persona_id))
-                            payload: dict = {"path": ""}
-                            if expressions:
-                                payload["expressions"] = expressions
-                            if animations:
-                                payload["animations"] = animations
-                            if char_anim_dir:
-                                payload["character_animations_dir"] = char_anim_dir
-                            await self.sio.emit(
-                                "avatar_load_model", payload
-                            )
-                            print(
-                                "[GUI] Avatar load_model emitted: (default)",
-                                flush=True,
-                            )
-                    except Exception as e:
-                        print(
-                            f"[GUI] Avatar load_model failed: {e}",
-                            flush=True,
-                        )
-            else:
-                current_state = "unknown"
-                if self._orchestrator._state_machine:
-                    current_state = self._orchestrator._state_machine.state.value
-                await self.sio.emit(
-                    "persona_change_failed",
-                    {"error": f"Cannot switch while {current_state}"},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def set_vad_config(sid: str, data: dict) -> None:
-            """Client updates VAD configuration."""
-            if self._orchestrator:
-                config = self._orchestrator._config
-                updated = False
-
-                if "threshold" in data:
-                    config.vad_threshold = float(data["threshold"])
-                    updated = True
-                if "min_speech_ms" in data:
-                    config.min_speech_ms = int(data["min_speech_ms"])
-                    updated = True
-                if "min_silence_ms" in data:
-                    config.min_silence_ms = int(data["min_silence_ms"])
-                    updated = True
-                if "speech_pad_ms" in data:
-                    config.speech_pad_ms = int(data["speech_pad_ms"])
-                    updated = True
-
-                if updated:
-                    # Propagate to live VAD instance
-                    sm = getattr(self._orchestrator, "_state_machine", None)
-                    if sm is not None:
-                        sm.update_vad_params(
-                            threshold=config.vad_threshold,
-                            min_speech_ms=config.min_speech_ms,
-                            min_silence_ms=config.min_silence_ms,
-                            speech_pad_ms=config.speech_pad_ms,
-                        )
-
-                    print(
-                        f"[GUI] VAD: threshold={config.vad_threshold:.2f}, "
-                        f"min_speech={config.min_speech_ms}ms, "
-                        f"min_silence={config.min_silence_ms}ms, "
-                        f"speech_pad={config.speech_pad_ms}ms",
-                        flush=True,
-                    )
-
-                    # Persist to YAML if config path is available
-                    persisted = False
-                    if self._config_path:
-                        try:
-                            config.save_to_yaml(self._config_path)
-                            persisted = True
-                            print(f"[GUI] VAD config persisted to {Path(self._config_path).name}", flush=True)
-                        except Exception as e:
-                            print(f"[GUI] Failed to persist VAD config: {e}", flush=True)
-
-                    # Emit updated config to confirm
-                    await self.sio.emit(
-                        "vad_config_updated",
-                        {
-                            "threshold": config.vad_threshold,
-                            "min_speech_ms": config.min_speech_ms,
-                            "min_silence_ms": config.min_silence_ms,
-                            "speech_pad_ms": config.speech_pad_ms,
-                            "persisted": persisted,
-                        },
-                        to=sid,
-                    )
-
-        @self.sio.event
-        async def set_pipeline_config(sid: str, data: dict) -> None:
-            """Client updates pipeline configuration."""
-            if self._orchestrator:
-                config = self._orchestrator._config
-                updated = False
-
-                if "summarization_threshold" in data:
-                    config.summarization_threshold = float(data["summarization_threshold"])
-                    updated = True
-
-                if updated:
-                    print(
-                        f"[GUI] Pipeline: summarization={config.summarization_threshold:.0%}",
-                        flush=True,
-                    )
-
-                    # Persist to YAML if config path is available
-                    persisted = False
-                    if self._config_path:
-                        try:
-                            config.save_to_yaml(self._config_path)
-                            persisted = True
-                            print(f"[GUI] Pipeline config persisted to {Path(self._config_path).name}", flush=True)
-                        except Exception as e:
-                            print(f"[GUI] Failed to persist Pipeline config: {e}", flush=True)
-
-                    await self.sio.emit(
-                        "pipeline_config_updated",
-                        {
-                            "summarization_threshold": config.summarization_threshold,
-                            "budget_strategy": config.budget_strategy,
-                            "persisted": persisted,
-                        },
-                        to=sid,
-                    )
-
-        @self.sio.event
-        async def set_memory_config(sid: str, data: dict) -> None:
-            """Client updates memory/RAG configuration."""
-            if self._orchestrator:
-                updated = False
-
-                # Use ellipsis as sentinel to distinguish "not provided" from "set to None"
-                top_k = None
-                relevance_threshold = ...
-                dedup_threshold = ...
-                reflection_interval = None
-                reflection_prompt = ...
-                reflection_system_message = ...
-                reflection_delimiter = None
-
-                if "top_k" in data:
-                    top_k = int(data["top_k"])
-                    updated = True
-                if "relevance_threshold" in data:
-                    val = data["relevance_threshold"]
-                    relevance_threshold = float(val) if val is not None else None
-                    updated = True
-                if "dedup_threshold" in data:
-                    val = data["dedup_threshold"]
-                    dedup_threshold = float(val) if val is not None else None
-                    updated = True
-                if "reflection_interval" in data:
-                    reflection_interval = int(data["reflection_interval"])
-                    updated = True
-                if "reflection_prompt" in data:
-                    val = data["reflection_prompt"]
-                    # Empty string or null → None (use built-in default)
-                    reflection_prompt = val if val else None
-                    updated = True
-                if "reflection_system_message" in data:
-                    val = data["reflection_system_message"]
-                    reflection_system_message = val if val else None
-                    updated = True
-                if "reflection_delimiter" in data:
-                    val = data["reflection_delimiter"]
-                    if val:  # Don't accept empty delimiter
-                        reflection_delimiter = str(val)
-                        updated = True
-
-                if updated:
-                    self._orchestrator.update_memory_config(
-                        top_k=top_k,
-                        relevance_threshold=relevance_threshold,
-                        dedup_threshold=dedup_threshold,
-                        reflection_interval=reflection_interval,
-                        reflection_prompt=reflection_prompt,
-                        reflection_system_message=reflection_system_message,
-                        reflection_delimiter=reflection_delimiter,
-                    )
-
-                    config = self._orchestrator._config
-                    print(
-                        f"[GUI] Memory: top_k={config.memory_config.rag_top_k}, "
-                        f"relevance_threshold={config.memory_config.relevance_threshold}, "
-                        f"dedup_threshold={config.memory_config.dedup_threshold}, "
-                        f"reflection_interval={config.memory_config.reflection_interval}",
-                        flush=True,
-                    )
-
-                    # Persist to YAML
-                    persisted = False
-                    if self._config_path:
-                        try:
-                            config.save_to_yaml(self._config_path)
-                            persisted = True
-                            print(f"[GUI] Memory config persisted to {Path(self._config_path).name}", flush=True)
-                        except Exception as e:
-                            print(f"[GUI] Failed to persist Memory config: {e}", flush=True)
-
-                    await self.sio.emit(
-                        "memory_config_updated",
-                        {
-                            "top_k": config.memory_config.rag_top_k,
-                            "relevance_threshold": config.memory_config.relevance_threshold,
-                            "dedup_threshold": config.memory_config.dedup_threshold,
-                            "reflection_interval": config.memory_config.reflection_interval,
-                            "reflection_prompt": config.memory_config.reflection_prompt,
-                            "reflection_system_message": config.memory_config.reflection_system_message,
-                            "reflection_delimiter": config.memory_config.reflection_delimiter,
-                            "enabled": config.memory_config.enabled,
-                            "persisted": persisted,
-                        },
-                        to=sid,
-                    )
-
-        # ============================================================
-        # NANO-102: Memory Curation Config — Socket Handler
-        # ============================================================
-
-        @self.sio.event
-        async def set_curation_config(sid: str, data: dict) -> None:
-            """Client updates memory curation configuration (NANO-102)."""
-            if self._orchestrator:
-                config = self._orchestrator._config
-                curation = config.memory_config.curation
-                updated = False
-
-                if "enabled" in data:
-                    curation.enabled = bool(data["enabled"])
-                    updated = True
-                if "api_key" in data:
-                    curation.api_key = data["api_key"]
-                    updated = True
-                if "model" in data:
-                    curation.model = str(data["model"])
-                    updated = True
-                if "prompt" in data:
-                    curation.prompt = data["prompt"]
-                    updated = True
-                if "timeout" in data:
-                    curation.timeout = float(data["timeout"])
-                    updated = True
-
-                if updated:
-                    print(
-                        f"[GUI] Curation: enabled={curation.enabled}, "
-                        f"model={curation.model}",
-                        flush=True,
-                    )
-
-                    # Persist to YAML
-                    persisted = False
-                    if self._config_path:
-                        try:
-                            config.save_to_yaml(self._config_path)
-                            persisted = True
-                            print(f"[GUI] Curation config persisted to {Path(self._config_path).name}", flush=True)
-                        except Exception as e:
-                            print(f"[GUI] Failed to persist Curation config: {e}", flush=True)
-
-                    await self.sio.emit(
-                        "curation_config_updated",
-                        {
-                            "enabled": curation.enabled,
-                            "api_key": curation.api_key,
-                            "model": curation.model,
-                            "prompt": curation.prompt,
-                            "timeout": curation.timeout,
-                            "persisted": persisted,
-                        },
-                        to=sid,
-                    )
-
-        # ============================================================
-        # NANO-045d: Prompt Injection Wrappers — Socket Handler
-        # ============================================================
-
-        @self.sio.event
-        async def set_prompt_config(sid: str, data: dict) -> None:
-            """Client updates prompt injection wrapper strings (NANO-045d)."""
-            if self._orchestrator:
-                # Use ellipsis sentinel for "not provided"
-                rag_prefix = ...
-                rag_suffix = ...
-                codex_prefix = ...
-                codex_suffix = ...
-                example_dialogue_prefix = ...
-                example_dialogue_suffix = ...
-                updated = False
-
-                if "rag_prefix" in data:
-                    rag_prefix = str(data["rag_prefix"])
-                    updated = True
-                if "rag_suffix" in data:
-                    rag_suffix = str(data["rag_suffix"])
-                    updated = True
-                if "codex_prefix" in data:
-                    codex_prefix = str(data["codex_prefix"])
-                    updated = True
-                if "codex_suffix" in data:
-                    codex_suffix = str(data["codex_suffix"])
-                    updated = True
-                if "example_dialogue_prefix" in data:
-                    example_dialogue_prefix = str(data["example_dialogue_prefix"])
-                    updated = True
-                if "example_dialogue_suffix" in data:
-                    example_dialogue_suffix = str(data["example_dialogue_suffix"])
-                    updated = True
-
-                if updated:
-                    self._orchestrator.update_prompt_config(
-                        rag_prefix=rag_prefix,
-                        rag_suffix=rag_suffix,
-                        codex_prefix=codex_prefix,
-                        codex_suffix=codex_suffix,
-                        example_dialogue_prefix=example_dialogue_prefix,
-                        example_dialogue_suffix=example_dialogue_suffix,
-                    )
-
-                    pc = self._orchestrator._config.prompt_config
-                    print(
-                        f"[GUI] Prompt wrappers updated: "
-                        f"rag_prefix={pc.rag_prefix[:40]}..., "
-                        f"codex_prefix={pc.codex_prefix[:40]}...",
-                        flush=True,
-                    )
-
-                    # Persist to YAML
-                    persisted = False
-                    if self._config_path:
-                        try:
-                            self._orchestrator._config.save_to_yaml(self._config_path)
-                            persisted = True
-                            print(f"[GUI] Prompt config persisted to {Path(self._config_path).name}", flush=True)
-                        except Exception as e:
-                            print(f"[GUI] Failed to persist prompt config: {e}", flush=True)
-
-                    await self.sio.emit(
-                        "prompt_config_updated",
-                        {
-                            "rag_prefix": pc.rag_prefix,
-                            "rag_suffix": pc.rag_suffix,
-                            "codex_prefix": pc.codex_prefix,
-                            "codex_suffix": pc.codex_suffix,
-                            "example_dialogue_prefix": pc.example_dialogue_prefix,
-                            "example_dialogue_suffix": pc.example_dialogue_suffix,
-                            "persisted": persisted,
-                        },
-                        to=sid,
-                    )
-
-        # ============================================================
-        # NANO-045c-1: Block Config — Socket Handlers
-        # ============================================================
-
-        @self.sio.event
-        async def request_block_config(sid: str, data: dict) -> None:
-            """Client requests current block configuration (NANO-045c-1, NANO-064b)."""
-            try:
-                if self._orchestrator:
-                    config = self._orchestrator.get_block_config()
-                else:
-                    config = self._get_block_config_pre_launch()
-                await self.sio.emit("block_config_loaded", config, to=sid)
-            except Exception as e:
-                print(f"[GUI] Error getting block config: {e}", flush=True)
-
-        @self.sio.event
-        async def set_block_config(sid: str, data: dict) -> None:
-            """Client updates block configuration (NANO-045c-1, NANO-064b)."""
-            try:
-                # Build config dict from provided fields
-                config: dict = {}
-                if "order" in data:
-                    config["order"] = list(data["order"])
-                if "disabled" in data:
-                    config["disabled"] = list(data["disabled"])
-                if "overrides" in data:
-                    config["overrides"] = dict(data["overrides"])
-
-                if self._orchestrator:
-                    # Live path: merge + hot-reload + YAML save
-                    existing = self._orchestrator._config.prompt_blocks
-                    if existing and isinstance(existing, dict):
-                        merged = {**existing, **config}
-                    else:
-                        merged = config
-
-                    self._orchestrator.update_block_config(merged)
-
-                    print(
-                        f"[GUI] Block config updated: "
-                        f"order={len(merged.get('order', []))} blocks, "
-                        f"disabled={len(merged.get('disabled', []))} blocks",
-                        flush=True,
-                    )
-
-                    persisted = False
-                    if self._config_path:
-                        try:
-                            self._orchestrator._config.save_block_config_to_yaml(
-                                self._config_path
-                            )
-                            persisted = True
-                            print(
-                                f"[GUI] Block config persisted to {Path(self._config_path).name}",
-                                flush=True,
-                            )
-                        except Exception as e:
-                            print(f"[GUI] Failed to persist block config: {e}", flush=True)
-
-                    current = self._orchestrator.get_block_config()
-                else:
-                    # Pre-launch path: YAML-only save (NANO-064b)
-                    existing = self._prompt_blocks_config
-                    if existing and isinstance(existing, dict):
-                        merged = {**existing, **config}
-                    else:
-                        merged = config
-
-                    self._prompt_blocks_config = merged
-                    persisted = self._save_block_config_pre_launch(merged)
-
-                    print(
-                        f"[GUI] Block config updated pre-launch: "
-                        f"order={len(merged.get('order', []))} blocks, "
-                        f"disabled={len(merged.get('disabled', []))} blocks",
-                        flush=True,
-                    )
-
-                    current = self._get_block_config_pre_launch()
-
-                await self.sio.emit(
-                    "block_config_updated",
-                    {
-                        "success": True,
-                        "persisted": persisted,
-                        "order": current["order"],
-                        "disabled": current["disabled"],
-                        "overrides": current["overrides"],
-                    },
-                    to=sid,
-                )
-            except Exception as e:
-                print(f"[GUI] Error updating block config: {e}", flush=True)
-                await self.sio.emit(
-                    "block_config_updated",
-                    {"success": False, "persisted": False, "order": [], "disabled": [], "overrides": {}},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def reset_block_config(sid: str, data: dict) -> None:
-            """Client resets block configuration to defaults (NANO-045c-1, NANO-064b)."""
-            try:
-                if self._orchestrator:
-                    self._orchestrator.reset_block_config()
-
-                    persisted = False
-                    if self._config_path:
-                        try:
-                            self._orchestrator._config.save_block_config_to_yaml(
-                                self._config_path
-                            )
-                            persisted = True
-                        except Exception as e:
-                            print(f"[GUI] Failed to persist block config reset: {e}", flush=True)
-
-                    current = self._orchestrator.get_block_config()
-                else:
-                    # Pre-launch: clear cached config, remove section from YAML
-                    self._prompt_blocks_config = None
-                    persisted = self._save_block_config_pre_launch({})
-                    current = self._get_block_config_pre_launch()
-
-                print("[GUI] Block config reset to defaults", flush=True)
-
-                await self.sio.emit(
-                    "block_config_updated",
-                    {
-                        "success": True,
-                        "persisted": persisted,
-                        "order": current["order"],
-                        "disabled": current["disabled"],
-                        "overrides": current["overrides"],
-                    },
-                    to=sid,
-                )
-            except Exception as e:
-                print(f"[GUI] Error resetting block config: {e}", flush=True)
-                await self.sio.emit(
-                    "block_config_updated",
-                    {"success": False, "persisted": False, "order": [], "disabled": [], "overrides": {}},
-                    to=sid,
-                )
-
-        # ============================================================
-        # NANO-053: Generation Parameters — Socket Handler
-        # ============================================================
-
-        @self.sio.event
-        async def set_generation_params(sid: str, data: dict) -> None:
-            """Client updates generation parameters at runtime (NANO-053, NANO-108)."""
-            if self._orchestrator:
-                temperature = ...
-                max_tokens = ...
-                top_p = ...
-                repeat_penalty = ...
-                repeat_last_n = ...
-                frequency_penalty = ...
-                presence_penalty = ...
-                updated = False
-
-                if "temperature" in data:
-                    val = float(data["temperature"])
-                    if 0.0 <= val <= 2.0:
-                        temperature = val
-                        updated = True
-                if "max_tokens" in data:
-                    val = int(data["max_tokens"])
-                    if 64 <= val <= 8192:
-                        max_tokens = val
-                        updated = True
-                if "top_p" in data:
-                    val = float(data["top_p"])
-                    if 0.0 <= val <= 1.0:
-                        top_p = val
-                        updated = True
-                if "repeat_penalty" in data:
-                    val = float(data["repeat_penalty"])
-                    if 0.0 <= val <= 2.0:
-                        repeat_penalty = val
-                        updated = True
-                if "repeat_last_n" in data:
-                    val = int(data["repeat_last_n"])
-                    if 0 <= val <= 2048:
-                        repeat_last_n = val
-                        updated = True
-                if "frequency_penalty" in data:
-                    val = float(data["frequency_penalty"])
-                    if -2.0 <= val <= 2.0:
-                        frequency_penalty = val
-                        updated = True
-                if "presence_penalty" in data:
-                    val = float(data["presence_penalty"])
-                    if -2.0 <= val <= 2.0:
-                        presence_penalty = val
-                        updated = True
-
-                if updated:
-                    self._orchestrator.update_generation_params(
-                        temperature=temperature,
-                        max_tokens=max_tokens,
-                        top_p=top_p,
-                        repeat_penalty=repeat_penalty,
-                        repeat_last_n=repeat_last_n,
-                        frequency_penalty=frequency_penalty,
-                        presence_penalty=presence_penalty,
-                    )
-
-                    pc = self._orchestrator._config.llm_config.provider_config
-                    print(
-                        f"[GUI] Generation params: "
-                        f"temp={pc.get('temperature', 0.7)}, "
-                        f"max_tokens={pc.get('max_tokens', 256)}, "
-                        f"top_p={pc.get('top_p', 0.95)}, "
-                        f"repeat_penalty={pc.get('repeat_penalty', 1.1)}, "
-                        f"repeat_last_n={pc.get('repeat_last_n', 64)}, "
-                        f"freq_penalty={pc.get('frequency_penalty', 0.0)}, "
-                        f"pres_penalty={pc.get('presence_penalty', 0.0)}",
-                        flush=True,
-                    )
-
-                    # Persist to YAML
-                    persisted = False
-                    if self._config_path:
-                        try:
-                            self._orchestrator._config.save_to_yaml(self._config_path)
-                            persisted = True
-                            print(f"[GUI] Generation params persisted to {Path(self._config_path).name}", flush=True)
-                        except Exception as e:
-                            print(f"[GUI] Failed to persist generation params: {e}", flush=True)
-
-                    await self.sio.emit(
-                        "generation_params_updated",
-                        {
-                            "temperature": pc.get("temperature", 0.7),
-                            "max_tokens": pc.get("max_tokens", 256),
-                            "top_p": pc.get("top_p", 0.95),
-                            "repeat_penalty": pc.get("repeat_penalty", 1.1),
-                            "repeat_last_n": pc.get("repeat_last_n", 64),
-                            "frequency_penalty": pc.get("frequency_penalty", 0.0),
-                            "presence_penalty": pc.get("presence_penalty", 0.0),
-                            "persisted": persisted,
-                        },
-                        to=sid,
-                    )
-
-        # ============================================================
-        # NANO-065a: Runtime Tools Toggle — Socket Handlers
-        # ============================================================
-
-        @self.sio.event
-        async def request_tools_config(sid: str, data: dict) -> None:
-            """Client requests current tools state (NANO-065a)."""
-            if self._orchestrator:
-                state = self._orchestrator.get_tools_state()
-            else:
-                # Pre-launch fallback: read from cached YAML config
-                cache = self._tools_config_cache or {}
-                master = cache.get("enabled", False)
-                tools_section = cache.get("tools", {})
-                tools = {}
-                for name, cfg in tools_section.items():
-                    label = name.replace("_", " ").title()
-                    tools[name] = {
-                        "enabled": cfg.get("enabled", True),
-                        "label": label,
-                    }
-                state = {"master_enabled": master, "tools": tools}
-
-            await self.sio.emit("tools_config_updated", state, to=sid)
-
-        @self.sio.event
-        async def set_tools_config(sid: str, data: dict) -> None:
-            """Client updates tools enable/disable state at runtime (NANO-065a)."""
-            if self._orchestrator:
-                master_enabled = ...
-                tools_changes = None
-
-                if "master_enabled" in data:
-                    master_enabled = bool(data["master_enabled"])
-
-                if "tools" in data and isinstance(data["tools"], dict):
-                    tools_changes = data["tools"]
-
-                result = self._orchestrator.update_tools_config(
-                    master_enabled=master_enabled,
-                    tools=tools_changes,
-                )
-
-                # NANO-089 Phase 4: relay errors to frontend
-                if not result.get("success", True):
-                    state = self._orchestrator.get_tools_state()
-                    state["error"] = result.get("error", "Tools toggle failed")
-                    await self.sio.emit("tools_config_updated", state, to=sid)
-                    return
-
-                print(
-                    f"[GUI] Tools config updated: "
-                    f"master={data.get('master_enabled', '(unchanged)')}, "
-                    f"tools={list(data.get('tools', {}).keys()) or '(unchanged)'}",
-                    flush=True,
-                )
-
-                # Persist to YAML
-                persisted = False
-                if self._config_path:
-                    try:
-                        self._orchestrator._config.save_to_yaml(self._config_path)
-                        persisted = True
-                        print(f"[GUI] Tools config persisted to {Path(self._config_path).name}", flush=True)
-                    except Exception as e:
-                        print(f"[GUI] Failed to persist tools config: {e}", flush=True)
-
-                # Get current state and emit
-                state = self._orchestrator.get_tools_state()
-                state["persisted"] = persisted
-                await self.sio.emit("tools_config_updated", state, to=sid)
-                await self._emit_health()
-
-        # ============================================================
-        # NANO-065b: Runtime LLM Provider/Model Swap — Socket Handlers
-        # ============================================================
-
-        @self.sio.event
-        async def request_llm_config(sid: str, data: dict) -> None:
-            """Client requests current LLM provider state (NANO-065b)."""
-            if self._orchestrator:
-                state = self._orchestrator.get_llm_state()
-            else:
-                # Pre-launch fallback: read from cached YAML config
-                cache = self._llm_config_cache or {}
-                state = {
-                    "provider": cache.get("provider", "llama"),
-                    "model": cache.get("model", ""),
-                    "context_size": cache.get("context_size"),
-                    "available_providers": list(
-                        (cache.get("providers") or {}).keys()
-                    ),
-                }
-
-            await self.sio.emit("llm_config_updated", state, to=sid)
-
-        @self.sio.event
-        async def set_llm_provider(sid: str, data: dict) -> None:
-            """Client requests LLM provider/model swap at runtime (NANO-065b)."""
-            if not self._orchestrator:
-                await self.sio.emit(
-                    "llm_config_updated",
-                    {"error": "Services not launched"},
-                    to=sid,
-                )
-                return
-
-            provider_name = data.get("provider")
-            provider_config = data.get("config")
-
-            if not provider_name or not isinstance(provider_config, dict):
-                await self.sio.emit(
-                    "llm_config_updated",
-                    {"error": "Missing provider name or config"},
-                    to=sid,
-                )
-                return
-
-            result = self._orchestrator.swap_llm_provider(
-                provider_name, provider_config
-            )
-
-            if not result.get("success"):
-                await self.sio.emit(
-                    "llm_config_updated",
-                    {"error": result.get("error", "Unknown error")},
-                    to=sid,
-                )
-                return
-
-            print(
-                f"[GUI] LLM provider swapped to {provider_name} "
-                f"(model: {result.get('model', 'unknown')})",
-                flush=True,
-            )
-
-            # Persist to YAML
-            persisted = False
-            if self._config_path:
-                try:
-                    self._orchestrator._config.save_to_yaml(self._config_path)
-                    persisted = True
-                    print(
-                        f"[GUI] LLM config persisted to "
-                        f"{Path(self._config_path).name}",
-                        flush=True,
-                    )
-                except Exception as e:
-                    print(f"[GUI] Failed to persist LLM config: {e}", flush=True)
-
-            result["persisted"] = persisted
-            await self.sio.emit("llm_config_updated", result, to=sid)
-            await self._emit_health()
-
-        @self.sio.event
-        async def request_local_llm_config(sid: str, data: dict) -> None:
-            """Return stored llama provider config for Dashboard hydration (NANO-065b)."""
-            if self._orchestrator:
-                llm_cfg = self._orchestrator._config.llm_config
-                llama_cfg = dict(llm_cfg.providers.get("llama", {}))
-            elif self._llm_config_cache:
-                llama_cfg = dict((self._llm_config_cache.get("providers") or {}).get("llama", {}))
-            else:
-                llama_cfg = {}
-
-            # Convert extra_args list → string for frontend display
-            if isinstance(llama_cfg.get("extra_args"), list):
-                llama_cfg["extra_args"] = " ".join(str(a) for a in llama_cfg["extra_args"])
-
-            # Convert tensor_split list → string for frontend display
-            if isinstance(llama_cfg.get("tensor_split"), list):
-                llama_cfg["tensor_split"] = ",".join(str(x) for x in llama_cfg["tensor_split"])
-
-            # Check if LLM service is currently running
-            server_running = False
-            if self._service_runner and self._service_runner.is_service_running("llm"):
-                server_running = True
-
-            await self.sio.emit(
-                "local_llm_config",
-                {"config": llama_cfg, "server_running": server_running},
-                to=sid,
-            )
-
-        @self.sio.event
-        async def launch_llm_server(sid: str, data: dict) -> None:
-            """Launch (or relaunch) a local llama-server from the Dashboard (NANO-065b)."""
-            from spindl.llm.builtin.llama.provider import LlamaProvider
-            from spindl.launcher.config import (
-                ServiceConfig, HealthCheckConfig, LLMProviderConfig,
-            )
-            from spindl.launcher.service_runner import ServiceRunner
-            from spindl.launcher.log_aggregator import LogAggregator
-
-            config = data.get("config", {})
-
-            # Convert extra_args string → list for backend (frontend sends string)
-            if isinstance(config.get("extra_args"), str):
-                raw = config["extra_args"].strip()
-                config["extra_args"] = raw.split() if raw else []
-
-            # Convert tensor_split string → list[float] for backend
-            if isinstance(config.get("tensor_split"), str):
-                raw_ts = config["tensor_split"].strip()
-                if raw_ts:
-                    config["tensor_split"] = [
-                        float(v.strip()) for v in raw_ts.split(",")
-                        if v.strip()
-                    ]
-                else:
-                    config.pop("tensor_split", None)
-
-            # Validate
-            errors = LlamaProvider.validate_config(config)
-            if errors:
-                await self.sio.emit(
-                    "llm_server_launched",
-                    {"success": False, "error": f"Validation failed: {'; '.join(errors)}"},
-                    to=sid,
-                )
-                return
-
-            host = config.get("host", "127.0.0.1")
-            port = config.get("port", 5557)
-
-            # NANO-079: Extract unified_vision flag before passing to provider
-            unified_vision = config.pop("unified_vision", False)
-
-            # Quick check: is a server already running at this address?
-            # Skip if unified_vision changed — the user is requesting a config
-            # change (e.g. adding mmproj + -np 2) that requires a real restart.
-            import requests
-            try:
-                resp = requests.get(f"http://{host}:{port}/health", timeout=3)
-                if resp.status_code == 200 and not unified_vision:
-                    await self.sio.emit(
-                        "llm_server_launched",
-                        {"success": True, "already_running": True},
-                        to=sid,
-                    )
-                    return
-            except Exception:
-                pass  # Not running — proceed with launch
-
-            # Build provider config
-            llm_provider_config = LLMProviderConfig(
-                provider="llama",
-                provider_config=config,
-            )
-
-            # NANO-079: When unified vision is active, set VLM provider config
-            # so the ServiceRunner injects -np 2 for dual-slot operation
-            vision_provider_config = None
-            if unified_vision:
-                from spindl.launcher.config import VisionProviderConfig
-                vision_provider_config = VisionProviderConfig(
-                    enabled=True,
-                    provider="llm",
-                    provider_config={},
-                )
-
-            context_size = config.get("context_size", 8192)
-
-            # Stop existing LLM service if running (model hot-swap)
-            if self._service_runner and self._service_runner.is_service_running("llm"):
-                # Distinguish cloud (just unregister) vs local (actual process stop)
-                is_cloud = "llm" in self._service_runner._cloud_services
-                if is_cloud:
-                    print("[GUI] Unregistering cloud LLM provider before local launch...", flush=True)
-                else:
-                    print("[GUI] Stopping existing LLM service for model swap...", flush=True)
-                self._service_runner.stop_service("llm")
-                self._launched_services.discard("llm")
-
-            # Create or update ServiceRunner
-            if self._service_runner is None:
-                if self._log_aggregator is None:
-                    self._log_aggregator = LogAggregator()
-                kwargs = dict(
-                    logger=self._log_aggregator,
-                    llm_provider_config=llm_provider_config,
-                    llm_context_size=context_size,
-                )
-                if vision_provider_config:
-                    kwargs["vision_provider_config"] = vision_provider_config
-                self._service_runner = ServiceRunner(**kwargs)
-            else:
-                # Update existing runner's LLM config + bust cache
-                self._service_runner._llm_provider_config = llm_provider_config
-                self._service_runner._llm_provider_class = None
-                self._service_runner._llm_context_size = context_size
-                # NANO-079: Update vision config for -np 2 injection
-                if vision_provider_config:
-                    self._service_runner._vision_provider_config = vision_provider_config
-
-            # Build synthetic ServiceConfig
-            svc_config = ServiceConfig(
-                name="llm",
-                platform="native",
-                command=None,
-                health_check=HealthCheckConfig(type="provider", timeout=90),
-            )
-
-            # Preview the command for diagnostics
-            try:
-                preview_cmd = self._service_runner._build_command(svc_config)
-                print(f"[GUI] LLM launch command: {preview_cmd}", flush=True)
-            except Exception as e:
-                print(f"[GUI] LLM command preview failed: {e}", flush=True)
-
-            # Launch in background thread
-            def _launch():
-                try:
-                    success = self._service_runner.start_service(svc_config)
-                except Exception as e:
-                    success = False
-                    print(f"[GUI] LLM launch exception: {e}", flush=True)
-
-                # Schedule async emit back on event loop
-                async def _emit_result():
-                    if success:
-                        self._launched_services.add("llm")
-                        print(
-                            f"[GUI] LLM server launched at {host}:{port}",
-                            flush=True,
-                        )
-
-                        # Persist config to YAML
-                        persisted = self._persist_local_llm_config(config)
-
-                        await self.sio.emit(
-                            "llm_server_launched",
-                            {"success": True, "persisted": persisted},
-                        )
-                    else:
-                        await self.sio.emit(
-                            "llm_server_launched",
-                            {
-                                "success": False,
-                                "error": f"LLM server failed to start at {host}:{port}",
-                            },
-                        )
-
-                loop = self._event_loop
-                if loop and loop.is_running():
-                    asyncio.run_coroutine_threadsafe(_emit_result(), loop)
-
-            thread = threading.Thread(target=_launch, daemon=True)
-            thread.start()
-
-            # Acknowledge receipt immediately
-            await self.sio.emit(
-                "llm_server_launched",
-                {"success": None, "status": "launching"},
-                to=sid,
-            )
-
-        @self.sio.event
-        async def request_openrouter_models(sid: str, data: dict) -> None:
-            """Fetch available models from OpenRouter API (NANO-065b)."""
-            import aiohttp
-
-            try:
-                # Resolve base URL from config
-                base_url = "https://openrouter.ai/api/v1"
-                if self._orchestrator:
-                    cfg = self._orchestrator._config.llm_config
-                    or_cfg = cfg.providers.get("openrouter", {})
-                    base_url = or_cfg.get("url", base_url).rstrip("/")
-                elif self._llm_config_cache:
-                    providers = self._llm_config_cache.get("providers", {})
-                    or_cfg = providers.get("openrouter", {})
-                    base_url = or_cfg.get("url", base_url).rstrip("/")
-
-                async with aiohttp.ClientSession() as session:
-                    async with session.get(
-                        f"{base_url}/models", timeout=aiohttp.ClientTimeout(total=10)
-                    ) as resp:
-                        if resp.status != 200:
-                            await self.sio.emit(
-                                "openrouter_models",
-                                {"error": f"OpenRouter API returned {resp.status}"},
-                                to=sid,
-                            )
-                            return
-                        body = await resp.json()
-
-                models = []
-                for m in body.get("data", []):
-                    models.append({
-                        "id": m.get("id", ""),
-                        "name": m.get("name", ""),
-                        "context_length": m.get("context_length"),
-                    })
-
-                # Sort by name for display
-                models.sort(key=lambda x: x["name"].lower())
-
-                await self.sio.emit(
-                    "openrouter_models", {"models": models}, to=sid
-                )
-
-            except Exception as e:
-                print(f"[GUI] Failed to fetch OpenRouter models: {e}", flush=True)
-                await self.sio.emit(
-                    "openrouter_models",
-                    {"error": f"Failed to fetch models: {e}"},
-                    to=sid,
-                )
-
-        # ============================================================
-        # NANO-065c: Runtime VLM Provider Swap — Socket Handlers
-        # ============================================================
-
-        @self.sio.event
-        async def request_vlm_config(sid: str, data: dict) -> None:
-            """Client requests current VLM provider state (NANO-065c)."""
-            if self._orchestrator:
-                state = self._orchestrator.get_vlm_state()
-            else:
-                # Pre-launch fallback: read from cached YAML config
-                cache = self._vlm_config_cache or {}
-                provider = cache.get("provider", "llama")
-                cloud_config = {}
-                if provider == "openai":
-                    stored = (cache.get("providers") or {}).get("openai", {})
-                    api_key = stored.get("api_key", "")
-                    if api_key and not api_key.startswith("${"):
-                        masked = api_key[:4] + "..." + api_key[-4:] if len(api_key) > 8 else "••••"
-                    else:
-                        masked = api_key
-                    cloud_config = {
-                        "api_key": masked,
-                        "model": stored.get("model", ""),
-                        "base_url": stored.get("base_url", ""),
-                    }
-                state = {
-                    "provider": provider,
-                    "available_providers": list(
-                        (cache.get("providers") or {}).keys()
-                    ) + ["llm"],
-                    "healthy": False,
-                }
-                if cloud_config:
-                    state["cloud_config"] = cloud_config
-
-            await self.sio.emit("vlm_config_updated", state, to=sid)
-
-        @self.sio.event
-        async def set_vlm_provider(sid: str, data: dict) -> None:
-            """Client requests VLM provider swap at runtime (NANO-065c)."""
-            if not self._orchestrator:
-                await self.sio.emit(
-                    "vlm_config_updated",
-                    {"error": "Services not launched"},
-                    to=sid,
-                )
-                return
-
-            provider_name = data.get("provider")
-            provider_config = data.get("config")
-
-            if not provider_name or not isinstance(provider_config, dict):
-                await self.sio.emit(
-                    "vlm_config_updated",
-                    {"error": "Missing provider name or config"},
-                    to=sid,
-                )
-                return
-
-            result = self._orchestrator.swap_vlm_provider(
-                provider_name, provider_config
-            )
-
-            if not result.get("success"):
-                await self.sio.emit(
-                    "vlm_config_updated",
-                    {"error": result.get("error", "Unknown error")},
-                    to=sid,
-                )
-                return
-
-            # Stop orphaned VLM server when switching to unified mode —
-            # the LLM handles vision now, no need for a dedicated VLM server
-            if provider_name == "llm" and self._service_runner:
-                if self._service_runner.is_service_running("vlm"):
-                    print("[GUI] Stopping dedicated VLM server (switching to unified mode)...", flush=True)
-                    self._service_runner.stop_service("vlm")
-                    self._launched_services.discard("vlm")
-
-            print(
-                f"[GUI] VLM provider swapped to {provider_name}",
-                flush=True,
-            )
-
-            # Persist to YAML
-            persisted = False
-            if self._config_path:
-                try:
-                    self._orchestrator._config.save_to_yaml(self._config_path)
-                    persisted = True
-                    print(
-                        f"[GUI] VLM config persisted to "
-                        f"{Path(self._config_path).name}",
-                        flush=True,
-                    )
-                except Exception as e:
-                    print(f"[GUI] Failed to persist VLM config: {e}", flush=True)
-
-            result["persisted"] = persisted
-            await self.sio.emit("vlm_config_updated", result, to=sid)
-            await self._emit_health()
-
-        # ============================================================
-        # NANO-079: Dashboard VLM Launch — Socket Handlers
-        # ============================================================
-
-        @self.sio.event
-        async def request_local_vlm_config(sid: str, data: dict) -> None:
-            """Client requests stored local VLM config for hydration (NANO-079)."""
-            vlm_cfg: dict = {}
-            if self._vlm_config_cache:
-                providers = self._vlm_config_cache.get("providers", {})
-                vlm_cfg = dict(providers.get("llama", {}))
-
-            # Convert extra_args list → string for frontend display
-            if isinstance(vlm_cfg.get("extra_args"), list):
-                vlm_cfg["extra_args"] = " ".join(str(a) for a in vlm_cfg["extra_args"])
-
-            # Convert tensor_split list → string for frontend display
-            if isinstance(vlm_cfg.get("tensor_split"), list):
-                vlm_cfg["tensor_split"] = ",".join(str(x) for x in vlm_cfg["tensor_split"])
-
-            # Check if VLM service is currently running
-            server_running = False
-            if self._service_runner and self._service_runner.is_service_running("vlm"):
-                server_running = True
-
-            await self.sio.emit(
-                "local_vlm_config",
-                {"config": vlm_cfg, "server_running": server_running},
-                to=sid,
-            )
-
-        @self.sio.event
-        async def launch_vlm_server(sid: str, data: dict) -> None:
-            """Launch (or relaunch) a local VLM llama-server from the Dashboard (NANO-079)."""
-            from spindl.vision.builtin.llama.provider import LlamaVLMProvider
-            from spindl.launcher.config import (
-                ServiceConfig, HealthCheckConfig, VisionProviderConfig,
-            )
-            from spindl.launcher.service_runner import ServiceRunner
-            from spindl.launcher.log_aggregator import LogAggregator
-
-            config = data.get("config", {})
-
-            # Convert extra_args string → list for backend (frontend sends string)
-            if isinstance(config.get("extra_args"), str):
-                raw = config["extra_args"].strip()
-                config["extra_args"] = raw.split() if raw else []
-
-            # Convert tensor_split string → list[float] for backend
-            if isinstance(config.get("tensor_split"), str):
-                raw_ts = config["tensor_split"].strip()
-                if raw_ts:
-                    config["tensor_split"] = [
-                        float(v.strip()) for v in raw_ts.split(",")
-                        if v.strip()
-                    ]
-                else:
-                    config.pop("tensor_split", None)
-
-            # Validate using VLM provider validator (not LLM)
-            errors = LlamaVLMProvider.validate_config(config)
-            if errors:
-                await self.sio.emit(
-                    "vlm_server_launched",
-                    {"success": False, "error": f"Validation failed: {'; '.join(errors)}"},
-                    to=sid,
-                )
-                return
-
-            host = config.get("host", "127.0.0.1")
-            port = config.get("port", 5558)
-
-            # Quick check: is a server already running at this address?
-            import requests
-            try:
-                resp = requests.get(f"http://{host}:{port}/health", timeout=3)
-                if resp.status_code == 200:
-                    await self.sio.emit(
-                        "vlm_server_launched",
-                        {"success": True, "already_running": True},
-                        to=sid,
-                    )
-                    return
-            except Exception:
-                pass  # Not running — proceed with launch
-
-            # Build vision provider config
-            vision_provider_config = VisionProviderConfig(
-                enabled=True,
-                provider="llama",
-                provider_config=config,
-            )
-
-            # Stop existing VLM service if running
-            if self._service_runner and self._service_runner.is_service_running("vlm"):
-                is_cloud = "vlm" in self._service_runner._cloud_services
-                if is_cloud:
-                    print("[GUI] Unregistering cloud VLM provider before local launch...", flush=True)
-                else:
-                    print("[GUI] Stopping existing VLM service for model swap...", flush=True)
-                self._service_runner.stop_service("vlm")
-                self._launched_services.discard("vlm")
-
-            # Create or update ServiceRunner
-            if self._service_runner is None:
-                if self._log_aggregator is None:
-                    self._log_aggregator = LogAggregator()
-                self._service_runner = ServiceRunner(
-                    logger=self._log_aggregator,
-                    vision_provider_config=vision_provider_config,
-                )
-            else:
-                # Update existing runner's VLM config + bust cache
-                self._service_runner._vision_provider_config = vision_provider_config
-                self._service_runner._vlm_provider_class = None
-
-            # Build synthetic ServiceConfig for VLM
-            svc_config = ServiceConfig(
-                name="vlm",
-                platform="native",
-                command=None,
-                health_check=HealthCheckConfig(type="provider", timeout=90),
-            )
-
-            # Preview the command for diagnostics
-            try:
-                preview_cmd = self._service_runner._build_command(svc_config)
-                print(f"[GUI] VLM launch command: {preview_cmd}", flush=True)
-            except Exception as e:
-                print(f"[GUI] VLM command preview failed: {e}", flush=True)
-
-            # Launch in background thread
-            def _launch():
-                try:
-                    success = self._service_runner.start_service(svc_config)
-                except Exception as e:
-                    success = False
-                    print(f"[GUI] VLM launch exception: {e}", flush=True)
-
-                # Schedule async emit back on event loop
-                async def _emit_result():
-                    if success:
-                        self._launched_services.add("vlm")
-                        print(
-                            f"[GUI] VLM server launched at {host}:{port}",
-                            flush=True,
-                        )
-
-                        # Persist config to YAML
-                        persisted = self._persist_local_vlm_config(config)
-
-                        await self.sio.emit(
-                            "vlm_server_launched",
-                            {"success": True, "persisted": persisted},
-                        )
-                    else:
-                        await self.sio.emit(
-                            "vlm_server_launched",
-                            {
-                                "success": False,
-                                "error": f"VLM server failed to start at {host}:{port}",
-                            },
-                        )
-
-                loop = self._event_loop
-                if loop and loop.is_running():
-                    asyncio.run_coroutine_threadsafe(_emit_result(), loop)
-
-            thread = threading.Thread(target=_launch, daemon=True)
-            thread.start()
-
-            # Acknowledge receipt immediately
-            await self.sio.emit(
-                "vlm_server_launched",
-                {"success": None, "status": "launching"},
-                to=sid,
-            )
-
-        # ============================================================
-        # NANO-056: Stimuli System — Socket Handlers
-        # ============================================================
-
-        @self.sio.event
-        async def set_stimuli_config(sid: str, data: dict) -> None:
-            """Client updates stimuli configuration at runtime (NANO-056)."""
-            if self._orchestrator:
-                enabled = data.get("enabled")
-                patience_enabled = data.get("patience_enabled")
-                patience_seconds = data.get("patience_seconds")
-                patience_prompt = data.get("patience_prompt")
-
-                # Twitch fields (NANO-056b)
-                twitch_enabled = data.get("twitch_enabled")
-                twitch_channel = data.get("twitch_channel")
-                twitch_app_id = data.get("twitch_app_id")
-                twitch_app_secret = data.get("twitch_app_secret")
-                twitch_buffer_size = data.get("twitch_buffer_size")
-                twitch_max_message_length = data.get("twitch_max_message_length")
-                twitch_prompt_template = data.get("twitch_prompt_template")
-
-                # Type coerce
-                if enabled is not None:
-                    enabled = bool(enabled)
-                if patience_enabled is not None:
-                    patience_enabled = bool(patience_enabled)
-                if patience_seconds is not None:
-                    patience_seconds = float(patience_seconds)
-                    if patience_seconds < 5:
-                        patience_seconds = 5.0
-                    elif patience_seconds > 600:
-                        patience_seconds = 600.0
-                if patience_prompt is not None:
-                    patience_prompt = str(patience_prompt).strip()
-                    if not patience_prompt:
-                        patience_prompt = None  # Don't set empty prompts
-
-                # Twitch type coercion (NANO-056b)
-                if twitch_enabled is not None:
-                    twitch_enabled = bool(twitch_enabled)
-                if twitch_channel is not None:
-                    twitch_channel = str(twitch_channel).strip()
-                if twitch_app_id is not None:
-                    twitch_app_id = str(twitch_app_id).strip()
-                if twitch_app_secret is not None:
-                    twitch_app_secret = str(twitch_app_secret).strip()
-                if twitch_buffer_size is not None:
-                    twitch_buffer_size = int(twitch_buffer_size)
-                    twitch_buffer_size = max(1, min(50, twitch_buffer_size))
-                if twitch_max_message_length is not None:
-                    twitch_max_message_length = int(twitch_max_message_length)
-                    twitch_max_message_length = max(50, min(1000, twitch_max_message_length))
-                if twitch_prompt_template is not None:
-                    twitch_prompt_template = str(twitch_prompt_template).strip()
-                    if not twitch_prompt_template:
-                        twitch_prompt_template = None
-
-                # Addressing-others contexts (NANO-110)
-                addressing_others_contexts = data.get("addressing_others_contexts")
-                if addressing_others_contexts is not None:
-                    if isinstance(addressing_others_contexts, list):
-                        # Validate: must have at least one entry
-                        if not addressing_others_contexts:
-                            addressing_others_contexts = None
-                        else:
-                            # Sanitize each context entry
-                            addressing_others_contexts = [
-                                {
-                                    "id": str(ctx.get("id", f"ctx_{i}")),
-                                    "label": str(ctx.get("label", "Others")).strip(),
-                                    "prompt": str(ctx.get("prompt", "")).strip(),
-                                }
-                                for i, ctx in enumerate(addressing_others_contexts)
-                                if isinstance(ctx, dict)
-                            ]
-                    else:
-                        addressing_others_contexts = None
-
-                self._orchestrator.update_stimuli_config(
-                    enabled=enabled,
-                    patience_enabled=patience_enabled,
-                    patience_seconds=patience_seconds,
-                    patience_prompt=patience_prompt,
-                    twitch_enabled=twitch_enabled,
-                    twitch_channel=twitch_channel,
-                    twitch_app_id=twitch_app_id,
-                    twitch_app_secret=twitch_app_secret,
-                    twitch_buffer_size=twitch_buffer_size,
-                    twitch_max_message_length=twitch_max_message_length,
-                    twitch_prompt_template=twitch_prompt_template,
-                    addressing_others_contexts=addressing_others_contexts,
-                )
-
-                cfg = self._orchestrator._config.stimuli_config
-                print(
-                    f"[GUI] Stimuli: enabled={cfg.enabled}, "
-                    f"patience_enabled={cfg.patience_enabled}, "
-                    f"patience_seconds={cfg.patience_seconds}, "
-                    f"twitch_enabled={cfg.twitch_enabled}",
-                    flush=True,
-                )
-
-                # Persist to YAML
-                persisted = False
-                if self._config_path:
-                    try:
-                        self._orchestrator._config.save_to_yaml(
-                            self._config_path
-                        )
-                        persisted = True
-                        print(
-                            f"[GUI] Stimuli config persisted to "
-                            f"{Path(self._config_path).name}",
-                            flush=True,
-                        )
-                    except Exception as e:
-                        print(
-                            f"[GUI] Failed to persist stimuli config: {e}",
-                            flush=True,
-                        )
-
-                stimuli_data = self._build_stimuli_hydration(cfg)
-                stimuli_data["persisted"] = persisted
-
-                # Broadcast to all clients (including Stream Deck) so
-                # context button grids rebuild when contexts are added/removed
-                await self.sio.emit(
-                    "stimuli_config_updated",
-                    stimuli_data,
-                )
-
-            if not self._orchestrator:
-                # Pre-launch: no persist, no emit. Credentials persist
-                # via test_twitch_credentials handler on Test Connection click.
-                # Orchestrator path handles persistence after launch.
-                pass
-
-        # ============================================================
-        # NANO-110: Tauri App Install — Socket Handlers
-        # ============================================================
-
-        @self.sio.event
-        async def check_tauri_install(sid: str, data: dict) -> None:
-            """Check which Tauri app binaries are installed (NANO-110)."""
-            project_root = Path(__file__).parent.parent.parent.parent
-            status = {
-                "avatar": self._tauri_binary_exists(project_root / "spindl-avatar"),
-                "subtitle": self._tauri_binary_exists(project_root / "spindl-subtitles"),
-                "stream_deck": self._tauri_binary_exists(project_root / "spindl-stream-deck"),
-            }
-            await self.sio.emit("tauri_install_status", status, to=sid)
-
-        @self.sio.event
-        async def install_tauri_apps(sid: str, data: dict) -> None:
-            """Trigger background build of all Tauri app binaries (NANO-110)."""
-            project_root = Path(__file__).parent.parent.parent.parent
-
-            apps = [
-                (project_root / "spindl-avatar", "Avatar"),
-                (project_root / "spindl-subtitles", "Subtitle"),
-                (project_root / "spindl-stream-deck", "Stream Deck"),
-            ]
-
-            missing = [
-                (d, n) for d, n in apps
-                if d.exists() and not self._tauri_binary_exists(d)
-            ]
-
-            if not missing:
-                await self.sio.emit("tauri_build_status", {
-                    "app": "all",
-                    "status": "ready",
-                    "message": "All overlay apps are already installed.",
-                    "progress": 0,
-                    "total": 0,
-                })
-                return
-
-            # Build all missing apps sequentially in one background thread.
-            # Shared workspace means first build compiles all deps,
-            # subsequent builds are near-instant (seconds).
-            self._tauri_install_all_in_background(missing)
-
-        # ============================================================
-        # NANO-110: Addressing Others — Socket Handlers
-        # ============================================================
-
-        @self.sio.event
-        async def addressing_others_start(sid: str, data: dict) -> None:
-            """Stream Deck / hotkey activates addressing-others mode (NANO-110)."""
-            print(f"[GUI] addressing_others_start received: {data}", flush=True)
-            if not self._orchestrator:
-                print("[GUI] addressing_others_start: no orchestrator", flush=True)
-                return
-            context_id = data.get("context_id", "ctx_0")
-            self._orchestrator.set_addressing_others(str(context_id))
-            await self.sio.emit(
-                "addressing_others_state",
-                {"active": True, "context_id": context_id},
-            )
-
-        @self.sio.event
-        async def addressing_others_stop(sid: str, data: dict) -> None:
-            """Stream Deck / hotkey deactivates addressing-others mode (NANO-110)."""
-            print(f"[GUI] addressing_others_stop received", flush=True)
-            if not self._orchestrator:
-                return
-            self._orchestrator.clear_addressing_others()
-            await self.sio.emit(
-                "addressing_others_state",
-                {"active": False, "context_id": None},
-            )
-
-        @self.sio.event
-        async def request_patience_progress(sid: str, data: dict) -> None:
-            """Client requests current PATIENCE progress (NANO-056)."""
-            if (
-                not self._orchestrator
-                or not self._orchestrator.stimuli_engine
-            ):
-                await self.sio.emit(
-                    "patience_progress",
-                    {"elapsed": 0, "total": 0, "progress": 0},
-                    to=sid,
-                )
-                return
-
-            engine = self._orchestrator.stimuli_engine
-            for module in engine.modules:
-                if module.name == "patience":
-                    progress_data = module.get_progress()
-                    progress_data["blocked"] = engine.is_blocked_by_playback or engine.is_blocked_by_typing
-                    progress_data["blocked_reason"] = (
-                        "typing" if engine.is_blocked_by_typing
-                        else "playback" if engine.is_blocked_by_playback
-                        else None
-                    )
-                    await self.sio.emit(
-                        "patience_progress",
-                        progress_data,
-                        to=sid,
-                    )
-                    return
-
-            # No PATIENCE module found
-            await self.sio.emit(
-                "patience_progress",
-                {"elapsed": 0, "total": 0, "progress": 0},
-                to=sid,
-            )
-
-        @self.sio.event
-        async def request_twitch_status(sid: str, data: dict) -> None:
-            """Client requests current Twitch module status (NANO-056b)."""
-            if (
-                not self._orchestrator
-                or not self._orchestrator.stimuli_engine
-            ):
-                await self.sio.emit(
-                    "twitch_status",
-                    {
-                        "connected": False,
-                        "channel": "",
-                        "buffer_count": 0,
-                        "recent_messages": [],
-                    },
-                    to=sid,
-                )
-                return
-
-            engine = self._orchestrator.stimuli_engine
-            for module in engine.modules:
-                if module.name == "twitch":
-                    await self.sio.emit(
-                        "twitch_status",
-                        {
-                            "connected": module.connected,
-                            "channel": module.channel,
-                            "buffer_count": module.buffer_count,
-                            "recent_messages": module.recent_messages,
-                        },
-                        to=sid,
-                    )
-                    return
-
-            # No Twitch module registered
-            await self.sio.emit(
-                "twitch_status",
-                {
-                    "connected": False,
-                    "channel": "",
-                    "buffer_count": 0,
-                    "recent_messages": [],
-                },
-                to=sid,
-            )
-
-        @self.sio.event
-        async def test_twitch_credentials(sid: str, data: dict) -> None:
-            """Validate Twitch app credentials without launching the module."""
-            import os
-            import re
-            env_pattern = re.compile(r"\$\{([^}]+)\}")
-
-            def resolve(value: str) -> str:
-                """Resolve ${ENV_VAR} patterns and fall back to env vars."""
-                value = value.strip()
-                if env_pattern.search(value):
-                    value = env_pattern.sub(
-                        lambda m: os.getenv(m.group(1), ""), value
-                    )
-                return value
-
-            app_id = resolve(str(data.get("app_id", ""))) or os.getenv("TWITCH_APP_ID", "")
-            app_secret = resolve(str(data.get("app_secret", ""))) or os.getenv("TWITCH_APP_SECRET", "")
-            channel = str(data.get("channel", "")).strip()
-
-            if not app_id or not app_secret:
-                await self.sio.emit(
-                    "twitch_credentials_result",
-                    {"success": False, "error": "App ID and App Secret are required."},
-                    to=sid,
-                )
-                return
-
-            try:
-                from twitchAPI.twitch import Twitch
-
-                twitch = await Twitch(app_id, app_secret)
-
-                # Validate channel exists if provided
-                if channel:
-                    user_list = [u async for u in twitch.get_users(logins=[channel])]
-                    if not user_list:
-                        await twitch.close()
-                        await self.sio.emit(
-                            "twitch_credentials_result",
-                            {"success": False, "error": f"Channel '{channel}' not found on Twitch."},
-                            to=sid,
-                        )
-                        return
-
-                await twitch.close()
-
-                # Persist validated credentials to YAML via regex surgery
-                if self._config_path:
-                    self._persist_twitch_credentials(
-                        data.get("channel", ""),
-                        data.get("app_id", ""),
-                        data.get("app_secret", ""),
-                    )
-
-                # Update stimuli cache (for next connect hydration)
-                # but do NOT emit stimuli_config_updated — that would
-                # overwrite the frontend's local input state with YAML values
-                if self._config_path:
-                    import yaml as _yaml
-                    with open(Path(self._config_path), "r", encoding="utf-8") as _f:
-                        _yd = _yaml.safe_load(_f) or {}
-                    from ..orchestrator.config import StimuliConfig
-                    _parsed = StimuliConfig.from_dict(_yd.get("stimuli", {}))
-                    self._stimuli_config_cache = GUIServer._build_stimuli_hydration(_parsed)
-
-                # If orchestrator is running, update the live module and bounce it
-                if self._orchestrator and self._orchestrator.stimuli_engine:
-                    cfg = self._orchestrator._config.stimuli_config
-                    cfg.twitch_channel = data.get("channel", "") or cfg.twitch_channel
-                    cfg.twitch_app_id = data.get("app_id", "") or cfg.twitch_app_id
-                    cfg.twitch_app_secret = data.get("app_secret", "") or cfg.twitch_app_secret
-
-                    engine = self._orchestrator.stimuli_engine
-                    for module in engine.modules:
-                        if module.name == "twitch":
-                            module.stop()
-                            module.channel = cfg.twitch_channel
-                            module.app_id = cfg.twitch_app_id
-                            module.app_secret = cfg.twitch_app_secret
-                            module.start()
-                            print("[GUI] Twitch module bounced with new credentials", flush=True)
-                            break
-
-                has_creds = bool(channel and app_id and app_secret)
-                await self.sio.emit(
-                    "twitch_credentials_result",
-                    {"success": True, "error": None, "has_credentials": has_creds},
-                    to=sid,
-                )
-                print(f"[GUI] Twitch credentials validated and persisted (channel={channel or 'N/A'})", flush=True)
-
-            except Exception as e:
-                error_msg = str(e)
-                if "401" in error_msg or "invalid" in error_msg.lower():
-                    error_msg = "Invalid App ID or App Secret."
-                await self.sio.emit(
-                    "twitch_credentials_result",
-                    {"success": False, "error": error_msg},
-                    to=sid,
-                )
-                print(f"[GUI] Twitch credential test failed: {e}", flush=True)
-
-        @self.sio.event
         async def typing_active(sid: str, data: dict) -> None:
             """Client signals user is typing (focus) or done typing (blur)."""
             if not self._orchestrator or not self._orchestrator.stimuli_engine:
                 return
             active = data.get("active", False)
             self._orchestrator.stimuli_engine.user_typing = active
-
-        # ============================================================
-        # NANO-060b: VTubeStudio — Socket Handlers
-        # ============================================================
-
-        @self.sio.event
-        async def set_vts_config(sid: str, data: dict) -> None:
-            """Client updates VTubeStudio configuration at runtime (NANO-060b)."""
-            if self._orchestrator:
-                enabled = data.get("enabled")
-                host = data.get("host")
-                port = data.get("port")
-
-                # Type coerce
-                if enabled is not None:
-                    enabled = bool(enabled)
-                if host is not None:
-                    host = str(host).strip()
-                    if not host:
-                        host = None
-                if port is not None:
-                    port = int(port)
-                    if port < 1 or port > 65535:
-                        port = None
-
-                self._orchestrator.update_vts_config(
-                    enabled=enabled,
-                    host=host,
-                    port=port,
-                )
-
-                cfg = self._orchestrator._config.vtubestudio_config
-                print(
-                    f"[GUI] VTS: enabled={cfg.enabled}, "
-                    f"host={cfg.host}, port={cfg.port}",
-                    flush=True,
-                )
-
-                # Persist to YAML
-                persisted = False
-                if self._config_path:
-                    try:
-                        self._orchestrator._config.save_to_yaml(
-                            self._config_path
-                        )
-                        persisted = True
-                        print(
-                            f"[GUI] VTS config persisted to "
-                            f"{Path(self._config_path).name}",
-                            flush=True,
-                        )
-                    except Exception as e:
-                        print(
-                            f"[GUI] Failed to persist VTS config: {e}",
-                            flush=True,
-                        )
-
-                await self.sio.emit(
-                    "vts_config_updated",
-                    {
-                        "enabled": cfg.enabled,
-                        "host": cfg.host,
-                        "port": cfg.port,
-                        "persisted": persisted,
-                    },
-                    to=sid,
-                )
-
-                # If enabling, schedule a delayed status push after driver connects
-                if enabled:
-                    async def _push_vts_status_after_connect(target_sid: str) -> None:
-                        """Poll driver status and push to client once connected."""
-                        import asyncio as _asyncio
-                        for _ in range(20):  # Up to 10 seconds
-                            await _asyncio.sleep(0.5)
-                            driver = (
-                                self._orchestrator.vts_driver
-                                if self._orchestrator
-                                else None
-                            )
-                            if driver and driver.is_connected():
-                                status = driver.get_status()
-                                status["enabled"] = True
-                                await self.sio.emit("vts_status", status, to=target_sid)
-                                return
-                    asyncio.ensure_future(_push_vts_status_after_connect(sid))
-
-        @self.sio.event
-        async def request_vts_status(sid: str, data: dict) -> None:
-            """Client requests VTubeStudio connection status (NANO-060b)."""
-            if (
-                not self._orchestrator
-                or not self._orchestrator.vts_driver
-            ):
-                cfg_enabled = False
-                if self._orchestrator:
-                    cfg_enabled = self._orchestrator._config.vtubestudio_config.enabled
-                await self.sio.emit(
-                    "vts_status",
-                    {
-                        "connected": False,
-                        "authenticated": False,
-                        "enabled": cfg_enabled,
-                        "model_name": None,
-                        "hotkeys": [],
-                        "expressions": [],
-                    },
-                    to=sid,
-                )
-                return
-
-            status = self._orchestrator.vts_driver.get_status()
-            status["enabled"] = self._orchestrator._config.vtubestudio_config.enabled
-            await self.sio.emit("vts_status", status, to=sid)
-
-        @self.sio.event
-        async def request_vts_hotkeys(sid: str, data: dict) -> None:
-            """Client requests available VTS hotkeys (NANO-060b).
-
-            Default: serve cached list for instant render.
-            With {refresh: true}: live query VTS then emit updated list.
-            """
-            driver = (
-                self._orchestrator.vts_driver
-                if self._orchestrator
-                else None
-            )
-            if not driver:
-                await self.sio.emit(
-                    "vts_hotkeys", {"hotkeys": []}, to=sid,
-                )
-                return
-
-            if data.get("refresh"):
-                # Live query — callback fires from driver thread
-                loop = self._event_loop or asyncio.get_event_loop()
-
-                def _on_hotkeys(result: list) -> None:
-                    asyncio.run_coroutine_threadsafe(
-                        self.sio.emit(
-                            "vts_hotkeys", {"hotkeys": result}, to=sid,
-                        ),
-                        loop,
-                    )
-
-                driver.request_hotkey_list(callback=_on_hotkeys)
-            else:
-                # Serve cached list
-                status = driver.get_status()
-                await self.sio.emit(
-                    "vts_hotkeys",
-                    {"hotkeys": status.get("hotkeys", [])},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def request_vts_expressions(sid: str, data: dict) -> None:
-            """Client requests available VTS expressions (NANO-060b).
-
-            Default: serve cached list for instant render.
-            With {refresh: true}: live query VTS then emit updated list.
-            """
-            driver = (
-                self._orchestrator.vts_driver
-                if self._orchestrator
-                else None
-            )
-            if not driver:
-                await self.sio.emit(
-                    "vts_expressions", {"expressions": []}, to=sid,
-                )
-                return
-
-            if data.get("refresh"):
-                # Live query — callback fires from driver thread
-                loop = self._event_loop or asyncio.get_event_loop()
-
-                def _on_expressions(result: list) -> None:
-                    asyncio.run_coroutine_threadsafe(
-                        self.sio.emit(
-                            "vts_expressions",
-                            {"expressions": result},
-                            to=sid,
-                        ),
-                        loop,
-                    )
-
-                driver.request_expression_list(callback=_on_expressions)
-            else:
-                # Serve cached list
-                status = driver.get_status()
-                await self.sio.emit(
-                    "vts_expressions",
-                    {"expressions": status.get("expressions", [])},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def send_vts_hotkey(sid: str, data: dict) -> None:
-            """Client triggers a VTS hotkey (NANO-060b)."""
-            driver = (
-                self._orchestrator.vts_driver
-                if self._orchestrator
-                else None
-            )
-            if not driver:
-                return
-
-            name = data.get("name")
-            if name:
-                driver.trigger_hotkey(str(name))
-                print(f"[GUI] VTS hotkey triggered: {name}", flush=True)
-                await self.sio.emit(
-                    "vts_hotkey_triggered",
-                    {"name": name},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def send_vts_expression(sid: str, data: dict) -> None:
-            """Client activates/deactivates a VTS expression (NANO-060b)."""
-            driver = (
-                self._orchestrator.vts_driver
-                if self._orchestrator
-                else None
-            )
-            if not driver:
-                return
-
-            file = data.get("file")
-            if file:
-                active = bool(data.get("active", True))
-                driver.trigger_expression(str(file), active)
-                print(
-                    f"[GUI] VTS expression {'activated' if active else 'deactivated'}: {file}",
-                    flush=True,
-                )
-                await self.sio.emit(
-                    "vts_expression_triggered",
-                    {"file": file, "active": active},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def send_vts_move(sid: str, data: dict) -> None:
-            """Client moves VTS model to a preset position (NANO-060b)."""
-            driver = (
-                self._orchestrator.vts_driver
-                if self._orchestrator
-                else None
-            )
-            if not driver:
-                return
-
-            preset = data.get("preset")
-            if preset:
-                driver.move_model(str(preset))
-                print(f"[GUI] VTS move: {preset}", flush=True)
-                await self.sio.emit(
-                    "vts_move_triggered",
-                    {"preset": preset},
-                    to=sid,
-                )
-
-        # ============================================================
-        # NANO-093: Avatar Config — Socket Handler
-        # ============================================================
-
-        @self.sio.event
-        async def set_avatar_config(sid: str, data: dict) -> None:
-            """Client updates avatar configuration at runtime (NANO-093, NANO-094)."""
-            if self._orchestrator:
-                config = self._orchestrator._config
-                updated = False
-
-                enabled = data.get("enabled")
-                emotion_classifier = data.get("emotion_classifier")
-                show_emotion_in_chat = data.get("show_emotion_in_chat")
-                emotion_confidence_threshold = data.get("emotion_confidence_threshold")
-                expression_fade_delay = data.get("expression_fade_delay")
-                subtitles_enabled = data.get("subtitles_enabled")
-                subtitle_fade_delay = data.get("subtitle_fade_delay")
-                avatar_always_on_top = data.get("avatar_always_on_top")
-                subtitle_always_on_top = data.get("subtitle_always_on_top")
-                stream_deck_enabled = data.get("stream_deck_enabled")
-
-                if enabled is not None:
-                    config.avatar_config.enabled = bool(enabled)
-                    updated = True
-                if emotion_classifier is not None:
-                    ec = str(emotion_classifier)
-                    if ec in ("classifier", "off"):
-                        config.avatar_config.emotion_classifier = ec
-                        updated = True
-                if show_emotion_in_chat is not None:
-                    config.avatar_config.show_emotion_in_chat = bool(show_emotion_in_chat)
-                    updated = True
-                if emotion_confidence_threshold is not None:
-                    try:
-                        threshold = float(emotion_confidence_threshold)
-                        if 0.0 <= threshold <= 1.0:
-                            config.avatar_config.emotion_confidence_threshold = threshold
-                            updated = True
-                    except (ValueError, TypeError):
-                        pass
-                if expression_fade_delay is not None:
-                    try:
-                        delay = float(expression_fade_delay)
-                        if 0.0 <= delay <= 10.0:
-                            config.avatar_config.expression_fade_delay = delay
-                            updated = True
-                    except (ValueError, TypeError):
-                        pass
-                if subtitles_enabled is not None:
-                    config.avatar_config.subtitles_enabled = bool(subtitles_enabled)
-                    updated = True
-                if subtitle_fade_delay is not None:
-                    try:
-                        delay = float(subtitle_fade_delay)
-                        if 0.0 <= delay <= 10.0:
-                            config.avatar_config.subtitle_fade_delay = delay
-                            updated = True
-                    except (ValueError, TypeError):
-                        pass
-                if avatar_always_on_top is not None:
-                    config.avatar_config.avatar_always_on_top = bool(avatar_always_on_top)
-                    updated = True
-                if subtitle_always_on_top is not None:
-                    config.avatar_config.subtitle_always_on_top = bool(subtitle_always_on_top)
-                    updated = True
-                if stream_deck_enabled is not None:
-                    config.avatar_config.stream_deck_enabled = bool(stream_deck_enabled)
-                    updated = True
-
-                if updated:
-                    cfg = config.avatar_config
-                    print(
-                        f"[GUI] Avatar: enabled={cfg.enabled}, "
-                        f"classifier={cfg.emotion_classifier}, "
-                        f"show_in_chat={cfg.show_emotion_in_chat}",
-                        flush=True,
-                    )
-
-                    persisted = False
-                    if self._config_path:
-                        try:
-                            config.save_to_yaml(self._config_path)
-                            persisted = True
-                            print(
-                                f"[GUI] Avatar config persisted to "
-                                f"{Path(self._config_path).name}",
-                                flush=True,
-                            )
-                        except Exception as e:
-                            print(
-                                f"[GUI] Failed to persist avatar config: {e}",
-                                flush=True,
-                            )
-
-                    await self.sio.emit(
-                        "avatar_config_updated",
-                        {
-                            "enabled": cfg.enabled,
-                            "emotion_classifier": cfg.emotion_classifier,
-                            "show_emotion_in_chat": cfg.show_emotion_in_chat,
-                            "emotion_confidence_threshold": cfg.emotion_confidence_threshold,
-                            "expression_fade_delay": cfg.expression_fade_delay,
-                            "subtitles_enabled": cfg.subtitles_enabled,
-                            "subtitle_fade_delay": cfg.subtitle_fade_delay,
-                            "avatar_always_on_top": cfg.avatar_always_on_top,
-                            "subtitle_always_on_top": cfg.subtitle_always_on_top,
-                            "stream_deck_enabled": cfg.stream_deck_enabled,
-                            "persisted": persisted,
-                        },
-                    )
-
-                # NANO-097: Process management on enabled transition
-                if enabled is not None:
-                    if bool(enabled):
-                        await self._avatar_spawn()
-                    else:
-                        self._avatar_kill()
-
-                # NANO-100: Subtitle process management on toggle
-                if subtitles_enabled is not None:
-                    if bool(subtitles_enabled):
-                        await self._subtitle_spawn()
-                    else:
-                        self._subtitle_kill()
-
-                # NANO-110: Stream Deck process management on toggle
-                if stream_deck_enabled is not None:
-                    if bool(stream_deck_enabled):
-                        await self._stream_deck_spawn()
-                    else:
-                        self._stream_deck_kill()
-
-            else:
-                # Pre-launch: orchestrator not running yet, but Tauri app
-                # builds and spawns don't need services. Handle toggle-only
-                # fields so users can build/launch windows from settings page
-                # before going through the launcher flow.
-                enabled = data.get("enabled")
-                subtitles_enabled = data.get("subtitles_enabled")
-                stream_deck_enabled = data.get("stream_deck_enabled")
-
-                if enabled is not None:
-                    if bool(enabled):
-                        await self._avatar_spawn()
-                    else:
-                        self._avatar_kill()
-                if subtitles_enabled is not None:
-                    if bool(subtitles_enabled):
-                        await self._subtitle_spawn()
-                    else:
-                        self._subtitle_kill()
-                if stream_deck_enabled is not None:
-                    if bool(stream_deck_enabled):
-                        await self._stream_deck_spawn()
-                    else:
-                        self._stream_deck_kill()
-
-        # ============================================================
-        # NANO-099: Avatar Rescan Animations — Socket Handler
-        # ============================================================
-
-        @self.sio.event
-        async def avatar_rescan_animations(sid: str) -> None:
-            """Notify avatar renderer to rescan animation files (NANO-099)."""
-            print("[GUI] Avatar rescan animations requested", flush=True)
-            # Reload base_animations from YAML into in-memory config
-            if self._orchestrator and self._config_path:
-                try:
-                    import yaml as _yaml
-                    with open(self._config_path, "r", encoding="utf-8") as f:
-                        raw = _yaml.safe_load(f)
-                    ba = (raw.get("avatar") or {}).get("base_animations") or {}
-                    self._orchestrator._config.avatar_config.base_animations = {
-                        "idle": ba.get("idle"),
-                        "happy": ba.get("happy"),
-                        "sad": ba.get("sad"),
-                    }
-                except Exception as e:
-                    print(f"[GUI] Failed to reload base_animations: {e}", flush=True)
-            # Forward to all connected clients with base_animations config
-            ba = {}
-            if self._orchestrator:
-                ba = self._orchestrator._config.avatar_config.base_animations
-            await self.sio.emit("avatar_rescan_animations", {"base_animations": ba})
-
-        # ============================================================
-        # NANO-042: Reasoning Config — Socket Handler
-        # ============================================================
-
-        @self.sio.event
-        async def set_reasoning_config(sid: str, data: dict) -> None:
-            """Client updates reasoning/thinking budget (persisted to YAML)."""
-            reasoning_budget = data.get("reasoning_budget")
-            if reasoning_budget is None:
-                return
-
-            reasoning_budget = int(reasoning_budget)
-            persisted = False
-
-            if self._config_path:
-                try:
-                    from spindl.orchestrator.config import _make_ruamel_yaml
-
-                    config_path = Path(self._config_path)
-                    ry = _make_ruamel_yaml()
-                    with open(config_path, "r", encoding="utf-8") as f:
-                        yaml_data = ry.load(f)
-
-                    llama = yaml_data.get("llm", {}).get("providers", {}).get("llama")
-                    if llama is not None:
-                        llama["reasoning_budget"] = reasoning_budget
-                        with open(config_path, "w", encoding="utf-8") as f:
-                            ry.dump(yaml_data, f)
-                        persisted = True
-
-                    print(
-                        f"[GUI] Reasoning: budget={reasoning_budget} persisted to {config_path.name}",
-                        flush=True,
-                    )
-                except Exception as e:
-                    print(f"[GUI] Failed to persist reasoning config: {e}", flush=True)
-
-            await self.sio.emit(
-                "reasoning_config_updated",
-                {
-                    "reasoning_budget": reasoning_budget,
-                    "persisted": persisted,
-                },
-                to=sid,
-            )
-
-        # ============================================================
-        # NANO-043 Phase 6: Memory Curation GUI — Socket Handlers
-        # ============================================================
-
-        @self.sio.event
-        async def request_memory_counts(sid: str, data: dict) -> None:
-            """Client requests memory collection counts."""
-            if not self._orchestrator or not self._orchestrator.memory_store:
-                await self.sio.emit(
-                    "memory_counts",
-                    {"global": 0, "general": 0, "flashcards": 0, "summaries": 0, "enabled": False},
-                    to=sid,
-                )
-                return
-
-            try:
-                counts = self._orchestrator.memory_store.counts
-                await self.sio.emit(
-                    "memory_counts",
-                    {**counts, "enabled": True},
-                    to=sid,
-                )
-            except Exception as e:
-                print(f"[GUI] Memory counts error: {e}", flush=True)
-                await self.sio.emit(
-                    "memory_counts",
-                    {"global": 0, "general": 0, "flashcards": 0, "summaries": 0, "enabled": True, "error": str(e)},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def request_memories(sid: str, data: dict) -> None:
-            """Client requests all memories from a specific collection."""
-            collection = data.get("collection")
-            if not collection or not self._orchestrator or not self._orchestrator.memory_store:
-                await self.sio.emit(
-                    "memory_list",
-                    {
-                        "collection": collection or "unknown",
-                        "memories": [],
-                        "error": "Memory system not available" if not (self._orchestrator and self._orchestrator.memory_store) else None,
-                    },
-                    to=sid,
-                )
-                return
-
-            try:
-                memories = self._orchestrator.memory_store.get_all(collection)
-                await self.sio.emit(
-                    "memory_list",
-                    {"collection": collection, "memories": memories},
-                    to=sid,
-                )
-            except Exception as e:
-                print(f"[GUI] Memory list error ({collection}): {e}", flush=True)
-                await self.sio.emit(
-                    "memory_list",
-                    {"collection": collection, "memories": [], "error": str(e)},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def add_general_memory(sid: str, data: dict) -> None:
-            """Client adds a new general memory."""
-            content = data.get("content", "").strip()
-            if not content:
-                await self.sio.emit(
-                    "memory_added",
-                    {"success": False, "error": "Content is required", "collection": "general"},
-                    to=sid,
-                )
-                return
-
-            if not self._orchestrator or not self._orchestrator.memory_store:
-                await self.sio.emit(
-                    "memory_added",
-                    {"success": False, "error": "Memory system not available", "collection": "general"},
-                    to=sid,
-                )
-                return
-
-            try:
-                metadata = {"type": "general", "source": "gui_manual"}
-                loop = asyncio.get_event_loop()
-                doc_id = await loop.run_in_executor(
-                    None,
-                    self._orchestrator.memory_store.add_general,
-                    content,
-                    metadata,
-                )
-                print(f"[GUI] General memory added: {doc_id}", flush=True)
-                await self.sio.emit(
-                    "memory_added",
-                    {
-                        "success": True,
-                        "collection": "general",
-                        "memory": {"id": doc_id, "content": content, "metadata": metadata},
-                    },
-                    to=sid,
-                )
-            except Exception as e:
-                print(f"[GUI] Add memory error: {e}", flush=True)
-                await self.sio.emit(
-                    "memory_added",
-                    {"success": False, "error": str(e), "collection": "general"},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def edit_general_memory(sid: str, data: dict) -> None:
-            """Client edits a general memory (delete + re-add with new embedding)."""
-            doc_id = data.get("id")
-            new_content = data.get("content", "").strip()
-
-            if not doc_id or not new_content:
-                await self.sio.emit(
-                    "memory_edited",
-                    {"success": False, "error": "ID and content are required"},
-                    to=sid,
-                )
-                return
-
-            if not self._orchestrator or not self._orchestrator.memory_store:
-                await self.sio.emit(
-                    "memory_edited",
-                    {"success": False, "error": "Memory system not available"},
-                    to=sid,
-                )
-                return
-
-            store = self._orchestrator.memory_store
-            try:
-                # Get original metadata before deleting
-                originals = store.get_all("general")
-                original = next((m for m in originals if m["id"] == doc_id), None)
-                original_meta = original["metadata"] if original else {}
-
-                # Delete old
-                store.delete("general", doc_id)
-
-                # Re-add with new content, preserving original timestamp
-                metadata = {
-                    **original_meta,
-                    "type": "general",
-                    "source": "gui_manual",
-                    "edited_at": time.strftime("%Y-%m-%dT%H:%M:%SZ", time.gmtime()),
-                }
-                loop = asyncio.get_event_loop()
-                new_id = await loop.run_in_executor(
-                    None,
-                    store.add_general,
-                    new_content,
-                    metadata,
-                )
-
-                print(f"[GUI] General memory edited: {doc_id} -> {new_id}", flush=True)
-                await self.sio.emit(
-                    "memory_edited",
-                    {
-                        "success": True,
-                        "old_id": doc_id,
-                        "memory": {"id": new_id, "content": new_content, "metadata": metadata},
-                    },
-                    to=sid,
-                )
-            except Exception as e:
-                print(f"[GUI] Edit memory error: {e}", flush=True)
-                await self.sio.emit(
-                    "memory_edited",
-                    {"success": False, "error": str(e)},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def add_global_memory(sid: str, data: dict) -> None:
-            """Client adds a new global memory (cross-character, NANO-105)."""
-            content = data.get("content", "").strip()
-            if not content:
-                await self.sio.emit(
-                    "memory_added",
-                    {"success": False, "error": "Content is required", "collection": "global"},
-                    to=sid,
-                )
-                return
-
-            if not self._orchestrator or not self._orchestrator.memory_store:
-                await self.sio.emit(
-                    "memory_added",
-                    {"success": False, "error": "Memory system not available", "collection": "global"},
-                    to=sid,
-                )
-                return
-
-            try:
-                metadata = {"type": "global", "source": "gui_manual"}
-                loop = asyncio.get_event_loop()
-                doc_id = await loop.run_in_executor(
-                    None,
-                    self._orchestrator.memory_store.add_global,
-                    content,
-                    metadata,
-                )
-                print(f"[GUI] Global memory added: {doc_id}", flush=True)
-                await self.sio.emit(
-                    "memory_added",
-                    {
-                        "success": True,
-                        "collection": "global",
-                        "memory": {"id": doc_id, "content": content, "metadata": metadata},
-                    },
-                    to=sid,
-                )
-            except Exception as e:
-                print(f"[GUI] Add global memory error: {e}", flush=True)
-                await self.sio.emit(
-                    "memory_added",
-                    {"success": False, "error": str(e), "collection": "global"},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def edit_global_memory(sid: str, data: dict) -> None:
-            """Client edits a global memory (NANO-105)."""
-            doc_id = data.get("id")
-            new_content = data.get("content", "").strip()
-
-            if not doc_id or not new_content:
-                await self.sio.emit(
-                    "memory_edited",
-                    {"success": False, "error": "ID and content are required"},
-                    to=sid,
-                )
-                return
-
-            if not self._orchestrator or not self._orchestrator.memory_store:
-                await self.sio.emit(
-                    "memory_edited",
-                    {"success": False, "error": "Memory system not available"},
-                    to=sid,
-                )
-                return
-
-            store = self._orchestrator.memory_store
-            try:
-                loop = asyncio.get_event_loop()
-                new_id = await loop.run_in_executor(
-                    None,
-                    store.edit_global,
-                    doc_id,
-                    new_content,
-                )
-
-                print(f"[GUI] Global memory edited: {doc_id} -> {new_id}", flush=True)
-                metadata = {"type": "global", "source": "gui_manual", "edited_at": time.strftime("%Y-%m-%dT%H:%M:%SZ", time.gmtime())}
-                await self.sio.emit(
-                    "memory_edited",
-                    {
-                        "success": True,
-                        "old_id": doc_id,
-                        "memory": {"id": new_id, "content": new_content, "metadata": metadata},
-                    },
-                    to=sid,
-                )
-            except Exception as e:
-                print(f"[GUI] Edit global memory error: {e}", flush=True)
-                await self.sio.emit(
-                    "memory_edited",
-                    {"success": False, "error": str(e)},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def delete_memory(sid: str, data: dict) -> None:
-            """Client deletes a memory from any collection."""
-            collection = data.get("collection")
-            doc_id = data.get("id")
-
-            if not collection or not doc_id:
-                await self.sio.emit(
-                    "memory_deleted",
-                    {"success": False, "error": "Collection and ID are required"},
-                    to=sid,
-                )
-                return
-
-            if not self._orchestrator or not self._orchestrator.memory_store:
-                await self.sio.emit(
-                    "memory_deleted",
-                    {"success": False, "error": "Memory system not available"},
-                    to=sid,
-                )
-                return
-
-            try:
-                success = self._orchestrator.memory_store.delete(collection, doc_id)
-                print(f"[GUI] Memory deleted: {collection}/{doc_id} (success={success})", flush=True)
-                await self.sio.emit(
-                    "memory_deleted",
-                    {"success": success, "collection": collection, "id": doc_id},
-                    to=sid,
-                )
-            except Exception as e:
-                print(f"[GUI] Delete memory error: {e}", flush=True)
-                await self.sio.emit(
-                    "memory_deleted",
-                    {"success": False, "error": str(e)},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def promote_memory(sid: str, data: dict) -> None:
-            """Promote a flash card or summary to general memory."""
-            source_collection = data.get("source_collection")
-            doc_id = data.get("id")
-            delete_source = data.get("delete_source", False)
-
-            if not source_collection or not doc_id:
-                await self.sio.emit(
-                    "memory_promoted",
-                    {"success": False, "error": "Source collection and ID required"},
-                    to=sid,
-                )
-                return
-
-            if source_collection not in ("flashcards", "summaries"):
-                await self.sio.emit(
-                    "memory_promoted",
-                    {"success": False, "error": "Can only promote from flashcards or summaries"},
-                    to=sid,
-                )
-                return
-
-            if not self._orchestrator or not self._orchestrator.memory_store:
-                await self.sio.emit(
-                    "memory_promoted",
-                    {"success": False, "error": "Memory system not available"},
-                    to=sid,
-                )
-                return
-
-            store = self._orchestrator.memory_store
-            try:
-                # Get the source document
-                all_docs = store.get_all(source_collection)
-                source_doc = next((m for m in all_docs if m["id"] == doc_id), None)
-                if not source_doc:
-                    await self.sio.emit(
-                        "memory_promoted",
-                        {"success": False, "error": "Source memory not found"},
-                        to=sid,
-                    )
-                    return
-
-                # Add to general with promotion metadata
-                metadata = {
-                    **source_doc.get("metadata", {}),
-                    "type": "general",
-                    "source": f"promoted_from_{source_collection}",
-                    "promoted_at": time.strftime("%Y-%m-%dT%H:%M:%SZ", time.gmtime()),
-                }
-                loop = asyncio.get_event_loop()
-                new_id = await loop.run_in_executor(
-                    None,
-                    store.add_general,
-                    source_doc["content"],
-                    metadata,
-                )
-
-                # Optionally delete from source
-                if delete_source:
-                    store.delete(source_collection, doc_id)
-
-                print(
-                    f"[GUI] Memory promoted: {source_collection}/{doc_id} -> general/{new_id}"
-                    f" (source {'deleted' if delete_source else 'kept'})",
-                    flush=True,
-                )
-                await self.sio.emit(
-                    "memory_promoted",
-                    {
-                        "success": True,
-                        "source_collection": source_collection,
-                        "source_id": doc_id,
-                        "new_id": new_id,
-                        "deleted_source": delete_source,
-                    },
-                    to=sid,
-                )
-            except Exception as e:
-                print(f"[GUI] Promote memory error: {e}", flush=True)
-                await self.sio.emit(
-                    "memory_promoted",
-                    {"success": False, "error": str(e)},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def search_memories(sid: str, data: dict) -> None:
-            """Semantic search across all memory collections."""
-            query_text = data.get("query", "").strip()
-            top_k = data.get("top_k", 10)
-
-            if not query_text:
-                await self.sio.emit(
-                    "memory_search_results",
-                    {"results": [], "query": ""},
-                    to=sid,
-                )
-                return
-
-            if not self._orchestrator or not self._orchestrator.memory_store:
-                await self.sio.emit(
-                    "memory_search_results",
-                    {"results": [], "query": query_text, "error": "Memory system not available"},
-                    to=sid,
-                )
-                return
-
-            try:
-                loop = asyncio.get_event_loop()
-                results = await loop.run_in_executor(
-                    None,
-                    self._orchestrator.memory_store.query,
-                    query_text,
-                    top_k,
-                )
-                await self.sio.emit(
-                    "memory_search_results",
-                    {"results": results, "query": query_text},
-                    to=sid,
-                )
-            except Exception as e:
-                print(f"[GUI] Memory search error: {e}", flush=True)
-                await self.sio.emit(
-                    "memory_search_results",
-                    {"results": [], "query": query_text, "error": str(e)},
-                    to=sid,
-                )
-
-        @self.sio.event
-        async def clear_flashcards(sid: str, data: dict) -> None:
-            """Clear all flash cards."""
-            if not self._orchestrator or not self._orchestrator.memory_store:
-                await self.sio.emit(
-                    "flashcards_cleared",
-                    {"success": False, "error": "Memory system not available"},
-                    to=sid,
-                )
-                return
-
-            try:
-                self._orchestrator.memory_store.clear_flash_cards()
-                print("[GUI] Flash cards cleared", flush=True)
-                await self.sio.emit(
-                    "flashcards_cleared",
-                    {"success": True},
-                    to=sid,
-                )
-            except Exception as e:
-                print(f"[GUI] Clear flashcards error: {e}", flush=True)
-                await self.sio.emit(
-                    "flashcards_cleared",
-                    {"success": False, "error": str(e)},
-                    to=sid,
-                )
-
-        # ============================================================
-        # End NANO-043 Phase 6 Memory Handlers
-        # ============================================================
 
         @self.sio.event
         async def pause_listening(sid: str) -> None:
@@ -4453,7 +1008,7 @@ class GUIServer:
                 "voice": persona.get("voice"),
             },
             "providers": {
-                "llm": self._get_llm_provider_info(config),
+                "llm": get_llm_provider_info(self, config),
                 "tts": {
                     "name": config.tts_config.provider,
                     "config": {},
@@ -4462,7 +1017,7 @@ class GUIServer:
                     "name": config.stt_config.provider,
                     "config": config.stt_config.provider_config,
                 },
-                "vlm": self._get_vlm_provider_info(config),
+                "vlm": get_vlm_provider_info(self, config),
                 "embedding": {
                     "base_url": config.memory_config.embedding_base_url,
                     "enabled": config.memory_config.enabled,
@@ -4512,7 +1067,7 @@ class GUIServer:
                     "frequency_penalty": config.llm_config.provider_config.get("frequency_penalty", 0.0),
                     "presence_penalty": config.llm_config.provider_config.get("presence_penalty", 0.0),
                 },
-                "stimuli": self._build_stimuli_hydration(config.stimuli_config),
+                "stimuli": build_stimuli_hydration(config.stimuli_config),
                 # NANO-065a: Tools runtime state
                 "tools": {
                     "master_enabled": config.tools_config.enabled,
@@ -4544,299 +1099,6 @@ class GUIServer:
             await self.sio.emit("config_loaded", config_data, to=sid)
         else:
             await self.sio.emit("config_loaded", config_data)
-
-    @staticmethod
-    def _extract_model_name(model_path: str) -> str:
-        """
-        Extract clean model name from model path.
-
-        Removes:
-        - Directory path (keeps filename only)
-        - .gguf extension
-        - Common quantization suffixes (Q8_0, Q4_K_M, IQ4_XS, etc.)
-
-        Args:
-            model_path: Full path to model file
-
-        Returns:
-            Clean model name for display
-        """
-        # Handle both forward and backslash paths
-        path = Path(model_path)
-        model_name = path.stem  # filename without extension
-
-        # Quantization suffixes to strip (ordered by specificity)
-        # Standard quantization: Q8_0, Q6_K, Q5_K_M, Q4_K_M, Q4_K_S, Q4_0, Q3_K_M, Q2_K
-        # Importance matrix: IQ4_XS, IQ4_NL, IQ3_XXS, IQ3_S, IQ2_XXS, IQ1_S
-        # Float formats: f16, f32, bf16
-        suffixes = [
-            # K-quants with size variants (most specific first)
-            "-Q6_K_L", "-Q5_K_L", "-Q4_K_L", "-Q3_K_L",
-            "-Q5_K_M", "-Q4_K_M", "-Q3_K_M", "-Q2_K_M",
-            "-Q5_K_S", "-Q4_K_S", "-Q3_K_S", "-Q2_K_S",
-            "-Q6_K", "-Q5_K", "-Q4_K", "-Q3_K", "-Q2_K",
-            # Standard quants
-            "-Q8_0", "-Q5_0", "-Q4_0", "-Q3_0", "-Q2_0",
-            "-Q5_1", "-Q4_1",
-            # Importance matrix quants
-            "-IQ4_XS", "-IQ4_NL", "-IQ3_XXS", "-IQ3_XS", "-IQ3_S", "-IQ3_M",
-            "-IQ2_XXS", "-IQ2_XS", "-IQ2_S", "-IQ2_M",
-            "-IQ1_S", "-IQ1_M",
-            # Float formats
-            "-f16", "-f32", "-bf16",
-        ]
-
-        for suffix in suffixes:
-            if model_name.endswith(suffix):
-                model_name = model_name[:-len(suffix)]
-                break
-
-        return model_name
-
-    def _persist_twitch_credentials(
-        self, channel: str, app_id: str, app_secret: str
-    ) -> None:
-        """Persist Twitch credentials to YAML via ruamel.yaml round-trip (NANO-106)."""
-        cfg_path = Path(self._config_path)
-        if not cfg_path.exists():
-            return
-
-        fields = {}
-        if channel:
-            fields["channel"] = channel
-        if app_id:
-            fields["app_id"] = app_id
-        if app_secret:
-            fields["app_secret"] = app_secret
-
-        if not fields:
-            return
-
-        from spindl.orchestrator.config import _make_ruamel_yaml
-
-        ry = _make_ruamel_yaml()
-        with open(cfg_path, "r", encoding="utf-8") as f:
-            data = ry.load(f)
-
-        if "stimuli" not in data:
-            data["stimuli"] = {}
-        if "twitch" not in data["stimuli"]:
-            data["stimuli"]["twitch"] = {}
-
-        tw = data["stimuli"]["twitch"]
-        for key, value in fields.items():
-            tw[key] = value
-
-        with open(cfg_path, "w", encoding="utf-8") as f:
-            ry.dump(data, f)
-
-        print(f"[GUI] Twitch credentials persisted to {cfg_path.name} "
-              f"(fields={set(fields.keys())})",
-              flush=True)
-
-    @staticmethod
-    def _build_stimuli_hydration(cfg) -> dict:
-        """Build stimuli config dict with resolved twitch_has_credentials flag."""
-        import os
-        import re
-        _env_pat = re.compile(r"\$\{([^}]+)\}")
-        def _resolve(v: str) -> str:
-            v = v or ""
-            if _env_pat.search(v):
-                v = _env_pat.sub(lambda m: os.getenv(m.group(1), ""), v)
-            return v
-        resolved_channel = _resolve(cfg.twitch_channel)
-        resolved_app_id = _resolve(cfg.twitch_app_id) or os.getenv("TWITCH_APP_ID", "")
-        resolved_app_secret = _resolve(cfg.twitch_app_secret) or os.getenv("TWITCH_APP_SECRET", "")
-
-        return {
-            "enabled": cfg.enabled,
-            "patience_enabled": cfg.patience_enabled,
-            "patience_seconds": cfg.patience_seconds,
-            "patience_prompt": cfg.patience_prompt,
-            "twitch_enabled": cfg.twitch_enabled,
-            "twitch_channel": cfg.twitch_channel or "",
-            "twitch_app_id": cfg.twitch_app_id or "",
-            "twitch_app_secret": cfg.twitch_app_secret or "",
-            "twitch_buffer_size": cfg.twitch_buffer_size,
-            "twitch_max_message_length": cfg.twitch_max_message_length,
-            "twitch_prompt_template": cfg.twitch_prompt_template,
-            "twitch_has_credentials": bool(
-                resolved_channel and resolved_app_id and resolved_app_secret
-            ),
-            # NANO-110: Addressing-others contexts
-            "addressing_others_contexts": [
-                {"id": ctx.id, "label": ctx.label, "prompt": ctx.prompt}
-                for ctx in cfg.addressing_others_contexts
-            ],
-        }
-
-    def _get_llm_provider_info(self, config) -> dict:
-        """Get LLM provider info, extracting model name for local/cloud providers."""
-        provider = config.llm_config.provider
-        provider_config = config.llm_config.provider_config
-
-        # Priority 1: Local providers (llama) - extract from model_path
-        if provider == "llama" and "model_path" in provider_config:
-            model_name = self._extract_model_name(provider_config["model_path"])
-            config_info: dict = {}
-            if provider_config.get("reasoning_format"):
-                config_info["reasoning_format"] = provider_config["reasoning_format"]
-            if "reasoning_budget" in provider_config:
-                config_info["reasoning_budget"] = provider_config["reasoning_budget"]
-            return {"name": model_name, "config": config_info}
-
-        # Priority 2: Cloud providers - check for explicit model field
-        if "model" in provider_config:
-            return {"name": provider_config["model"], "config": {}}
-
-        # Priority 3: Fallback to provider name
-        return {"name": provider, "config": {}}
-
-    def _get_vlm_provider_info(self, config) -> dict:
-        """Get VLM provider info, extracting model name for local/cloud providers."""
-        provider = config.vlm_config.provider
-        provider_config = config.vlm_config.providers.get(provider, {})
-
-        # Priority 1: Local providers (llama) - extract from model_path
-        if provider == "llama" and "model_path" in provider_config:
-            model_name = self._extract_model_name(provider_config["model_path"])
-            return {"name": model_name, "config": {}}
-
-        # Priority 2: Cloud providers - check for explicit model field
-        if "model" in provider_config:
-            return {"name": provider_config["model"], "config": {}}
-
-        # Priority 3: Fallback to provider name
-        return {"name": provider, "config": {}}
-
-    async def _emit_personas(self, sid: Optional[str] = None) -> None:
-        """Emit list of available personas."""
-        # NANO-048: Resolve personas_dir from orchestrator (preferred) or pre-launch config
-        personas_dir_str = None
-        if self._orchestrator:
-            personas_dir_str = self._orchestrator._config.personas_dir
-        elif self._personas_dir:
-            personas_dir_str = self._personas_dir
-
-        if not personas_dir_str:
-            return
-
-        personas_dir = Path(personas_dir_str)
-        personas = []
-        active_persona = self._orchestrator.persona.get("id", "unknown") if self._orchestrator and self._orchestrator.persona else "unknown"
-
-        if personas_dir.exists():
-            import yaml
-
-            for filepath in sorted(personas_dir.glob("*.yaml")):
-                try:
-                    with open(filepath, "r", encoding="utf-8") as f:
-                        data = yaml.safe_load(f)
-                        if data:
-                            personas.append({
-                                "id": filepath.stem,
-                                "name": data.get("name", filepath.stem),
-                                "description": data.get("description", ""),
-                            })
-                except Exception as e:
-                    print(f"[GUI] Failed to read persona {filepath.name}: {e}", flush=True)
-
-        if sid:
-            await self.sio.emit(
-                "persona_list",
-                {"personas": personas, "active": active_persona},
-                to=sid,
-            )
-        else:
-            await self.sio.emit(
-                "persona_list",
-                {"personas": personas, "active": active_persona},
-            )
-
-    async def _emit_sessions(
-        self, sid: Optional[str] = None, persona_filter: Optional[str] = None
-    ) -> None:
-        """Emit list of conversation sessions."""
-        from spindl.history import jsonl_store
-
-        # NANO-048: Resolve conversations_dir from orchestrator (preferred) or pre-launch config
-        conv_dir_str = None
-        if self._orchestrator:
-            conv_dir_str = self._orchestrator._config.conversations_dir
-        elif self._conversations_dir:
-            conv_dir_str = self._conversations_dir
-
-        if not conv_dir_str:
-            return
-
-        conversations_dir = Path(conv_dir_str)
-        sessions = []
-
-        if conversations_dir.exists():
-            for filepath in sorted(
-                (p for p in conversations_dir.glob("*.jsonl") if ".snapshot." not in p.name),
-                key=lambda p: p.stat().st_mtime, reverse=True,
-            ):
-                # Parse filename: {persona}_{timestamp}.jsonl
-                parts = filepath.stem.rsplit("_", 2)
-                if len(parts) >= 2:
-                    persona = parts[0]
-                    if persona_filter and persona != persona_filter:
-                        continue
-
-                    stat = filepath.stat()
-
-                    # Count turns (read file to get actual count)
-                    try:
-                        turns = jsonl_store.read_turns(filepath)
-                        turn_count = len(turns)
-                        visible_count = len([t for t in turns if not t.get("hidden", False)])
-                    except Exception:
-                        turn_count = 0
-                        visible_count = 0
-
-                    sessions.append({
-                        "filepath": str(filepath),
-                        "persona": persona,
-                        "timestamp": filepath.stem.split("_", 1)[1] if "_" in filepath.stem else "",
-                        "turn_count": turn_count,
-                        "visible_count": visible_count,
-                        "file_size": stat.st_size,
-                    })
-
-        # Include the orchestrator's current session so the frontend can badge it
-        active_session = None
-        if self._orchestrator and self._orchestrator.session_file:
-            active_session = str(self._orchestrator.session_file)
-
-        payload = {"sessions": sessions, "active_session": active_session}
-        if sid:
-            await self.sio.emit("session_list", payload, to=sid)
-        else:
-            await self.sio.emit("session_list", payload)
-
-    async def _emit_session_detail(self, sid: str, filepath: str) -> None:
-        """Emit detailed session data."""
-        from spindl.history import jsonl_store
-
-        try:
-            path = Path(filepath)
-            if not path.exists():
-                await self.sio.emit("error", {"message": "Session not found"}, to=sid)
-                return
-
-            turns = jsonl_store.read_turns(path)
-            await self.sio.emit(
-                "session_detail",
-                {"filepath": filepath, "turns": turns},
-                to=sid,
-            )
-        except Exception as e:
-            print(f"[GUI] Error reading session: {e}", flush=True)
-            await self.sio.emit("error", {"message": str(e)}, to=sid)
-
-    # === Public emit methods for EventBridge ===
 
     async def emit_state_changed(
         self, from_state: str, to_state: str, trigger: str, timestamp: str
@@ -5078,250 +1340,6 @@ class GUIServer:
     # NANO-097: Avatar Process Management
     # ============================================================
 
-    def _tauri_binary_path(self, app_dir: Path) -> Optional[Path]:
-        """Return the Tauri app binary path if it exists, else None."""
-        import platform
-        ext = ".exe" if platform.system() == "Windows" else ""
-        cargo_name = app_dir.name
-        project_root = app_dir.parent
-        # Check release first (installed via Install button), then debug (dev builds)
-        for profile in ("release", "debug"):
-            workspace = project_root / "target" / profile / f"{cargo_name}{ext}"
-            if workspace.exists():
-                return workspace
-            legacy = app_dir / "src-tauri" / "target" / profile / f"{cargo_name}{ext}"
-            if legacy.exists():
-                return legacy
-        return None
-
-    def _tauri_binary_exists(self, app_dir: Path) -> bool:
-        """Check if a Tauri app binary exists (workspace or legacy path)."""
-        return self._tauri_binary_path(app_dir) is not None
-
-    def _tauri_install_in_background(self, app_dir: Path, app_name: str) -> None:
-        """
-        Build a Tauri app binary in a background thread (NANO-110).
-
-        Streams cargo compilation progress to the frontend via socket events.
-        Non-blocking — returns immediately, build runs in a daemon thread.
-        On completion, emits 'ready' or 'failed' status.
-        """
-        import asyncio
-        import platform
-        import re
-        import threading
-
-        ext = ".exe" if platform.system() == "Windows" else ""
-        cargo_name = app_dir.name
-        project_root = app_dir.parent
-        app_key = app_name.lower().replace(" ", "_")
-
-        def _emit(status: str, message: str, progress: int = 0, total: int = 0) -> None:
-            if self._event_loop and self.sio:
-                asyncio.run_coroutine_threadsafe(
-                    self.sio.emit("tauri_build_status", {
-                        "app": app_key,
-                        "status": status,
-                        "message": message,
-                        "progress": progress,
-                        "total": total,
-                    }),
-                    self._event_loop,
-                )
-
-        def _build() -> None:
-            _emit("building", f"Installing {app_name} (first time only — this may take a few minutes)...")
-            print(f"[GUI] {app_name} binary not found — building (first-time only)...", flush=True)
-
-            building_re = re.compile(r"\[=*>?\s*\]\s+(\d+)/(\d+)")
-            compiling_count = 0
-
-            try:
-                proc = subprocess.Popen(
-                    ["cargo", "build", "--release", "-p", cargo_name],
-                    cwd=str(project_root),
-                    stdout=subprocess.PIPE,
-                    stderr=subprocess.STDOUT,
-                    text=True,
-                    bufsize=1,
-                )
-
-                for line in iter(proc.stdout.readline, ""):
-                    line = line.strip()
-                    if not line:
-                        continue
-
-                    m = building_re.search(line)
-                    if m:
-                        current, total = int(m.group(1)), int(m.group(2))
-                        _emit("building", f"Compiling crate {current}/{total}...", current, total)
-                    elif "Compiling" in line:
-                        compiling_count += 1
-                        parts = line.split("Compiling")
-                        crate = parts[-1].strip().split(" ")[0] if len(parts) > 1 else "..."
-                        _emit("building", f"Compiling {crate}...", compiling_count, 0)
-
-                    print(f"[GUI] {app_name} build: {line}", flush=True)
-
-                proc.stdout.close()
-                ret = proc.wait(timeout=600)
-
-                if ret != 0:
-                    print(f"[GUI] {app_name} cargo build failed (exit {ret})", flush=True)
-                    _emit("failed", f"{app_name} build failed.")
-                    return
-
-                print(f"[GUI] {app_name} build complete", flush=True)
-                _emit("ready", f"{app_name} installed successfully.")
-
-            except subprocess.TimeoutExpired:
-                print(f"[GUI] {app_name} cargo build timed out", flush=True)
-                _emit("failed", f"{app_name} build timed out.")
-                try:
-                    proc.kill()
-                except Exception:
-                    pass
-            except Exception as e:
-                print(f"[GUI] {app_name} cargo build error: {e}", flush=True)
-                _emit("failed", f"{app_name} build error: {e}")
-
-        thread = threading.Thread(target=_build, daemon=True, name=f"tauri-build-{app_key}")
-        thread.start()
-
-    def _tauri_install_all_in_background(self, apps: list[tuple[Path, str]]) -> None:
-        """
-        Build multiple Tauri apps sequentially in a background thread (NANO-110).
-
-        First app compiles all shared deps (~2-3 min). Subsequent apps reuse
-        the cached deps and compile in seconds. Emits 'ready' only after ALL
-        apps are built.
-        """
-        import asyncio
-        import platform
-        import re
-        import threading
-
-        ext = ".exe" if platform.system() == "Windows" else ""
-
-        def _emit(status: str, message: str, progress: int = 0, total: int = 0) -> None:
-            if self._event_loop and self.sio:
-                asyncio.run_coroutine_threadsafe(
-                    self.sio.emit("tauri_build_status", {
-                        "app": "all",
-                        "status": status,
-                        "message": message,
-                        "progress": progress,
-                        "total": total,
-                    }),
-                    self._event_loop,
-                )
-
-        def _build_all() -> None:
-            building_re = re.compile(r"\[=*>?\s*\]\s+(\d+)/(\d+)")
-
-            npm = "npm.cmd" if platform.system() == "Windows" else "npm"
-
-            for i, (app_dir, app_name) in enumerate(apps):
-                cargo_name = app_dir.name
-                project_root = app_dir.parent
-
-                label = f"({i + 1}/{len(apps)}) {app_name}"
-
-                # Step 1: Build frontend dist (vite build — skip tsc for resilience)
-                dist_dir = app_dir / "dist"
-                npx = "npx.cmd" if platform.system() == "Windows" else "npx"
-                if not dist_dir.exists():
-                    _emit("building", f"{label}: building frontend...")
-                    print(f"[GUI] {app_name}: building frontend dist...", flush=True)
-                    try:
-                        # Ensure node_modules exist
-                        node_modules = app_dir / "node_modules"
-                        if not node_modules.exists():
-                            subprocess.run(
-                                [npm, "install"],
-                                cwd=str(app_dir),
-                                stdout=subprocess.DEVNULL,
-                                stderr=subprocess.DEVNULL,
-                                timeout=120,
-                            )
-                        # vite build directly — more resilient than tsc + vite
-                        vite_result = subprocess.run(
-                            [npx, "vite", "build"],
-                            cwd=str(app_dir),
-                            stdout=subprocess.DEVNULL,
-                            stderr=subprocess.DEVNULL,
-                            timeout=120,
-                        )
-                        if vite_result.returncode != 0:
-                            print(f"[GUI] {app_name} frontend build failed", flush=True)
-                            _emit("failed", f"{app_name} frontend build failed.")
-                            return
-                    except Exception as e:
-                        print(f"[GUI] {app_name} frontend build error: {e}", flush=True)
-                        _emit("failed", f"{app_name} frontend build error.")
-                        return
-
-                # Step 2: Build Rust binary
-                _emit("building", f"Installing {label}...")
-                print(f"[GUI] Building {label}...", flush=True)
-
-                compiling_count = 0
-                try:
-                    proc = subprocess.Popen(
-                        ["cargo", "build", "--release", "-p", cargo_name],
-                        cwd=str(project_root),
-                        stdout=subprocess.PIPE,
-                        stderr=subprocess.STDOUT,
-                        text=True,
-                        bufsize=1,
-                    )
-
-                    for line in iter(proc.stdout.readline, ""):
-                        line = line.strip()
-                        if not line:
-                            continue
-
-                        m = building_re.search(line)
-                        if m:
-                            current, total = int(m.group(1)), int(m.group(2))
-                            _emit("building", f"{label}: crate {current}/{total}", current, total)
-                        elif "Compiling" in line:
-                            compiling_count += 1
-                            parts = line.split("Compiling")
-                            crate = parts[-1].strip().split(" ")[0] if len(parts) > 1 else "..."
-                            _emit("building", f"{label}: {crate}", compiling_count, 0)
-
-                        print(f"[GUI] {app_name} build: {line}", flush=True)
-
-                    proc.stdout.close()
-                    ret = proc.wait(timeout=600)
-
-                    if ret != 0:
-                        print(f"[GUI] {app_name} build failed (exit {ret})", flush=True)
-                        _emit("failed", f"{app_name} build failed.")
-                        return
-
-                    print(f"[GUI] {app_name} build complete", flush=True)
-
-                except subprocess.TimeoutExpired:
-                    print(f"[GUI] {app_name} build timed out", flush=True)
-                    _emit("failed", f"{app_name} build timed out.")
-                    try:
-                        proc.kill()
-                    except Exception:
-                        pass
-                    return
-                except Exception as e:
-                    print(f"[GUI] {app_name} build error: {e}", flush=True)
-                    _emit("failed", f"{app_name} build error: {e}")
-                    return
-
-            # All done
-            _emit("ready", "All overlay apps installed successfully.")
-            print("[GUI] All Tauri apps built successfully", flush=True)
-
-        thread = threading.Thread(target=_build_all, daemon=True, name="tauri-install-all")
-        thread.start()
 
     async def _avatar_spawn(self) -> None:
         """Spawn the avatar renderer if no avatar client is already connected."""
@@ -5345,7 +1363,7 @@ class GUIServer:
             return
 
         # Check binary exists — install must happen first via Install button
-        if not self._tauri_binary_exists(avatar_dir):
+        if not tauri_binary_exists(avatar_dir):
             print("[GUI] Avatar binary not installed — use Install button", flush=True)
             return
 
@@ -5414,7 +1432,7 @@ class GUIServer:
             return
 
         # Check binary exists — install must happen first via Install button
-        if not self._tauri_binary_exists(subtitle_dir):
+        if not tauri_binary_exists(subtitle_dir):
             print("[GUI] Subtitle binary not installed — use Install button", flush=True)
             return
 
@@ -5482,7 +1500,7 @@ class GUIServer:
             return
 
         # Check binary exists — install must happen first via Install button
-        binary = self._tauri_binary_path(deck_dir)
+        binary = tauri_binary_path(deck_dir)
         if not binary:
             print("[GUI] Stream Deck binary not installed — use Install button", flush=True)
             return
